\name{FPI2002Data}
\alias{FPI2002Data}
\docType{data}
\title{
A real life dataset consisting of the bilateral investment between many different countries in 2002, as well as the corresponding covariates from 2001. This was to eliminate any dual causality element in data analysis.
}
\description{
The dataset is a merging of CEPII CPIS data, the CEPII "Network Trade" dataset and the "Life During Growth" dataset from the world bank.
}
\usage{data("FPI2002Data")}
\format{
  A data frame with 1868 observations on the following 46 variables.
  \describe{
    \item{\code{iso_d}}{a character vector}
    \item{\code{iso_o}}{a character vector}
    \item{\code{year}}{a numeric vector}
    \item{\code{contig}}{a numeric vector}
    \item{\code{comlang_off}}{a numeric vector}
    \item{\code{distw}}{a numeric vector}
    \item{\code{pop_o}}{a numeric vector}
    \item{\code{gdp_o}}{a numeric vector}
    \item{\code{gdpcap_o}}{a numeric vector}
    \item{\code{iso2_o}}{a character vector}
    \item{\code{pop_d}}{a numeric vector}
    \item{\code{gdp_d}}{a numeric vector}
    \item{\code{gdpcap_d}}{a numeric vector}
    \item{\code{iso2_d}}{a character vector}
    \item{\code{heg_d}}{a numeric vector}
    \item{\code{conflict}}{a numeric vector}
    \item{\code{indepdate}}{a numeric vector}
    \item{\code{heg_o}}{a numeric vector}
    \item{\code{col_to}}{a numeric vector}
    \item{\code{col_fr}}{a numeric vector}
    \item{\code{col_hist}}{a numeric vector}
    \item{\code{col_cur}}{a numeric vector}
    \item{\code{sever}}{a numeric vector}
    \item{\code{sib_conflict}}{a numeric vector}
    \item{\code{gatt_o}}{a numeric vector}
    \item{\code{gatt_d}}{a numeric vector}
    \item{\code{rta}}{a numeric vector}
    \item{\code{comleg}}{a numeric vector}
    \item{\code{comcur}}{a numeric vector}
    \item{\code{acp_to_eu}}{a numeric vector}
    \item{\code{gsp}}{a numeric vector}
    \item{\code{eu_to_acp}}{a numeric vector}
    \item{\code{gsp_rec}}{a numeric vector}
    \item{\code{flow}}{a numeric vector}
    \item{\code{validmirror}}{a numeric vector}
    \item{\code{family}}{a character vector}
    \item{\code{FPIflow}}{a factor with levels \code{-} \code{-15} \code{-2} \code{-26} \code{-49} \code{-9} \code{...} \code{0} \code{1} \code{1,001} \code{1,002} \code{1,006} \code{1,012} \code{1,018} \code{1,028} \code{1,031} \code{1,038} \code{1,039} \code{1,046} \code{1,051} \code{1,060} \code{1,064} \code{1,070} \code{1,081} \code{1,083} \code{1,099} \code{1,100} \code{1,103} \code{1,111} \code{1,114} \code{1,116} \code{1,120} \code{1,124} \code{1,128} \code{1,130} \code{1,138} \code{1,143} \code{1,145} \code{1,147} \code{1,152} \code{1,154} \code{1,156} \code{1,161} \code{1,167} \code{1,167,313} \code{1,169} \code{1,172} \code{1,179} \code{1,185} \code{1,195} \code{1,196} \code{1,203} \code{1,205} \code{1,206} \code{1,208} \code{1,214} \code{1,234} \code{1,238} \code{1,239} \code{1,241} \code{1,251} \code{1,253} \code{1,257} \code{1,265} \code{1,267} \code{1,281,806} \code{1,282} \code{1,283} \code{1,284} \code{1,286} \code{1,288} \code{1,289,749} \code{1,289,876} \code{1,297} \code{1,298} \code{1,300} \code{1,304} \code{1,304,044} \code{1,307} \code{1,313} \code{1,314} \code{1,315} \code{1,321} \code{1,322} \code{1,325} \code{1,331} \code{1,333} \code{1,347} \code{1,359} \code{1,364} \code{1,367} \code{1,371} \code{1,372} \code{1,376} \code{1,377} \code{1,385} \code{1,389} \code{1,400} \code{1,403} \code{1,406} \code{1,407} \code{1,411} \code{1,415} \code{1,416} \code{1,421} \code{1,433} \code{1,435} \code{1,436} \code{1,437} \code{1,441} \code{1,444} \code{1,445} \code{1,446} \code{1,463} \code{1,473} \code{1,482} \code{1,505} \code{1,526} \code{1,528} \code{1,529} \code{1,530} \code{1,534} \code{1,537} \code{1,543} \code{1,555} \code{1,560} \code{1,582} \code{1,586} \code{1,591} \code{1,592} \code{1,605} \code{1,607} \code{1,615} \code{1,619} \code{1,622} \code{1,628} \code{1,629} \code{1,630} \code{1,632} \code{1,642} \code{1,643} \code{1,646} \code{1,647} \code{1,652} \code{1,667} \code{1,669} \code{1,673} \code{1,695} \code{1,697} \code{1,706} \code{1,707} \code{1,716} \code{1,726} \code{1,737} \code{1,749} \code{1,752} \code{1,763} \code{1,773} \code{1,778} \code{1,789} \code{1,814} \code{1,820} \code{1,829} \code{1,836} \code{1,841} \code{1,844} \code{1,846} \code{1,852} \code{1,858} \code{1,867} \code{1,868} \code{1,874} \code{1,878} \code{1,879} \code{1,884} \code{1,887} \code{1,897} \code{1,901} \code{1,903} \code{1,906} \code{1,920} \code{1,934} \code{1,957} \code{1,980} \code{10} \code{10,208} \code{10,335} \code{10,389} \code{10,399} \code{10,428} \code{10,542} \code{10,637} \code{10,694} \code{10,705} \code{10,754} \code{10,785} \code{10,802} \code{10,969} \code{10,992} \code{10,993} \code{100} \code{100,757} \code{101} \code{102} \code{103} \code{103,235} \code{103,660} \code{104} \code{105} \code{106} \code{106,052} \code{107} \code{108} \code{108,168} \code{109} \code{11} \code{11,054} \code{11,123} \code{11,143} \code{11,368} \code{11,508} \code{11,534} \code{11,543} \code{11,645} \code{11,663} \code{11,853} \code{11,881} \code{11,883} \code{11,961} \code{110} \code{110,356} \code{111} \code{111,307} \code{112} \code{112,285} \code{112,457} \code{113} \code{114} \code{115} \code{116} \code{116,530} \code{117} \code{117,085} \code{118} \code{119} \code{119,217} \code{12} \code{12,062} \code{12,115} \code{12,117} \code{12,121} \code{12,206} \code{12,289} \code{12,307} \code{12,319} \code{12,360} \code{12,422} \code{12,672} \code{12,719,412} \code{12,728} \code{12,916} \code{12,933} \code{12,947} \code{12,955} \code{12,987} \code{120} \code{122} \code{123} \code{124} \code{125} \code{125,478} \code{126} \code{127} \code{128} \code{129} \code{13} \code{13,035} \code{13,182} \code{13,489} \code{13,554} \code{13,702} \code{13,897} \code{130} \code{130,030} \code{130,577} \code{131} \code{132} \code{132,785} \code{133} \code{134} \code{134,142} \code{135} \code{136} \code{138} \code{14} \code{14,066} \code{14,140} \code{14,456} \code{14,512} \code{14,572} \code{14,576} \code{14,641} \code{14,872} \code{140} \code{141} \code{142} \code{142,411} \code{143} \code{143,572} \code{144} \code{145} \code{145,557} \code{145,558} \code{145,827} \code{146} \code{146,431} \code{147} \code{148} \code{149} \code{149,082} \code{15} \code{15,013} \code{15,132} \code{15,184} \code{15,280} \code{15,330} \code{15,521} \code{15,623} \code{15,662} \code{15,924} \code{150} \code{150,101} \code{151} \code{152} \code{152,701} \code{153} \code{153,568} \code{154} \code{155} \code{155,710} \code{156} \code{157} \code{158} \code{159} \code{16} \code{16,026} \code{16,046} \code{16,363} \code{16,397} \code{16,484} \code{16,585} \code{16,676} \code{16,679} \code{16,774} \code{16,820} \code{16,861} \code{16,901} \code{16,999} \code{160} \code{161} \code{162} \code{162,433} \code{163} \code{164} \code{165} \code{166} \code{167} \code{168} \code{168,501} \code{169} \code{17} \code{17,133} \code{17,273} \code{17,608} \code{17,776} \code{17,779} \code{170} \code{170,828} \code{171} \code{173} \code{173,143} \code{174} \code{174,551} \code{175} \code{176} \code{177} \code{177,597} \code{177,910} \code{178} \code{179} \code{18} \code{18,125} \code{18,336} \code{18,455} \code{18,528} \code{18,575} \code{18,639} \code{18,647} \code{18,804} \code{180} \code{182} \code{183} \code{184} \code{185} \code{186} \code{187} \code{189} \code{19} \code{19,180} \code{19,246} \code{19,352} \code{19,705} \code{19,813} \code{19,860} \code{19,895} \code{191} \code{192} \code{193} \code{194} \code{196} \code{197,839} \code{198} \code{199} \code{2} \code{2,001} \code{2,011} \code{2,016} \code{2,026} \code{2,041} \code{2,052} \code{2,057} \code{2,067} \code{2,071} \code{2,074} \code{2,075} \code{2,078} \code{2,094} \code{2,095} \code{2,104} \code{2,114} \code{2,118} \code{2,123} \code{2,132} \code{2,133} \code{2,135} \code{2,140} \code{2,146} \code{2,168} \code{2,178} \code{2,184} \code{2,185} \code{2,189} \code{2,196} \code{2,206} \code{2,219} \code{2,230} \code{2,232} \code{2,234} \code{2,235} \code{2,247} \code{2,267} \code{2,277} \code{2,279} \code{2,281} \code{2,287} \code{2,290} \code{2,297} \code{2,303} \code{2,303,603} \code{2,309} \code{2,318} \code{2,331} \code{2,341} \code{2,360} \code{2,364} \code{2,369} \code{2,378} \code{2,389} \code{2,400} \code{2,403} \code{2,407} \code{2,415} \code{2,421} \code{2,423} \code{2,425} \code{2,433} \code{2,435} \code{2,438} \code{2,440} \code{2,448} \code{2,450} \code{2,466} \code{2,469} \code{2,471} \code{2,485} \code{2,492} \code{2,531} \code{2,532} \code{2,533} \code{2,538} \code{2,610} \code{2,625} \code{2,626} \code{2,645} \code{2,646} \code{2,663} \code{2,685} \code{2,688} \code{2,690} \code{2,693} \code{2,696} \code{2,698} \code{2,721} \code{2,723} \code{2,724} \code{2,737} \code{2,739} \code{2,741} \code{2,760} \code{2,762} \code{2,765} \code{2,769} \code{2,770} \code{2,816} \code{2,822} \code{2,842} \code{2,843} \code{2,847} \code{2,862} \code{2,893} \code{2,903} \code{2,907} \code{2,919} \code{2,929} \code{2,933} \code{2,975} \code{20} \code{20,230} \code{20,417} \code{20,421} \code{20,887} \code{201} \code{202} \code{203} \code{204} \code{205} \code{205,600} \code{206} \code{207} \code{208} \code{209} \code{21} \code{21,117} \code{21,233} \code{21,489} \code{21,905} \code{21,910} \code{210} \code{211} \code{211,444} \code{212} \code{213} \code{214} \code{215} \code{217} \code{218} \code{218,587} \code{219} \code{22} \code{22,039} \code{22,551} \code{22,709} \code{22,755} \code{22,818} \code{22,972} \code{220} \code{221} \code{222} \code{223} \code{224} \code{225} \code{228} \code{23} \code{23,466} \code{23,752} \code{23,763} \code{230} \code{231} \code{232} \code{233} \code{234} \code{235} \code{236} \code{238} \code{24} \code{24,013} \code{24,255} \code{24,711} \code{240} \code{241} \code{242} \code{243} \code{245} \code{246} \code{247} \code{248} \code{249} \code{25} \code{25,008} \code{25,154} \code{25,247} \code{25,395} \code{250} \code{251} \code{252} \code{254} \code{257} \code{258} \code{259} \code{26} \code{26,137} \code{26,254} \code{26,657} \code{26,821} \code{26,850} \code{260} \code{261} \code{261,240} \code{263} \code{264} \code{265} \code{266} \code{267} \code{268} \code{269} \code{27} \code{27,342} \code{270} \code{271} \code{273} \code{274} \code{275} \code{277} \code{278} \code{279} \code{279,013} \code{28} \code{28,374} \code{28,401} \code{28,712} \code{28,947} \code{280} \code{280,426} \code{281} \code{283} \code{285} \code{285,067} \code{286} \code{287} \code{288} \code{289} \code{29} \code{29,195} \code{29,411} \code{29,741} \code{29,983} \code{290} \code{291} \code{292} \code{293} \code{295} \code{296} \code{298} \code{299} \code{3} \code{3,002} \code{3,004} \code{3,022} \code{3,048} \code{3,052} \code{3,055} \code{3,071} \code{3,085} \code{3,094} \code{3,099} \code{3,105,841} \code{3,110} \code{3,128} \code{3,146} \code{3,147} \code{3,205} \code{3,208} \code{3,217} \code{3,228} \code{3,251} \code{3,268} \code{3,280} \code{3,286} \code{3,295} \code{3,309} \code{3,316} \code{3,325} \code{3,330} \code{3,348} \code{3,445} \code{3,447} \code{3,460} \code{3,533} \code{3,536} \code{3,539} \code{3,544} \code{3,547} \code{3,564} \code{3,594} \code{3,616} \code{3,643} \code{3,646} \code{3,655} \code{3,656} \code{3,660} \code{3,674} \code{3,677} \code{3,690} \code{3,728} \code{3,742} \code{3,756} \code{3,764} \code{3,766} \code{3,798} \code{3,808} \code{3,815} \code{3,854} \code{3,869} \code{3,875} \code{3,891} \code{3,961} \code{3,963} \code{30} \code{30,260} \code{30,311} \code{30,472} \code{30,519} \code{30,643} \code{300} \code{301} \code{304} \code{306} \code{307} \code{308} \code{308,986} \code{309} \code{31} \code{31,070} \code{310} \code{312} \code{313} \code{314} \code{315} \code{317} \code{318} \code{319} \code{319,795} \code{32} \code{32,047} \code{32,304} \code{32,431} \code{32,783} \code{32,974} \code{320} \code{321} \code{322} \code{325} \code{326} \code{327} \code{328} \code{329} \code{33} \code{33,333} \code{33,731} \code{33,986} \code{330} \code{332} \code{333} \code{334} \code{335} \code{336} \code{338} \code{34} \code{34,130} \code{34,475} \code{342} \code{343} \code{345} \code{348} \code{35} \code{350} \code{352} \code{353} \code{354} \code{355} \code{356} \code{357} \code{358} \code{359} \code{36} \code{36,830} \code{36,932} \code{361} \code{362} \code{364} \code{368} \code{369} \code{37} \code{37,787} \code{370} \code{372} \code{374} \code{375} \code{376} \code{377} \code{379} \code{38} \code{382} \code{384} \code{385} \code{387} \code{388} \code{389} \code{39} \code{39,228} \code{39,253} \code{39,874} \code{39,882} \code{39,922} \code{390} \code{391} \code{393} \code{395} \code{396} \code{397} \code{4} \code{4,015} \code{4,022} \code{4,056} \code{4,063} \code{4,077} \code{4,089} \code{4,129} \code{4,139} \code{4,147} \code{4,161} \code{4,170} \code{4,189} \code{4,258} \code{4,304} \code{4,336} \code{4,346} \code{4,358} \code{4,385} \code{4,398} \code{4,410} \code{4,506} \code{4,509} \code{4,513} \code{4,522} \code{4,524} \code{4,533} \code{4,563} \code{4,607} \code{4,615} \code{4,625} \code{4,645} \code{4,649} \code{4,673} \code{4,676} \code{4,713} \code{4,730} \code{4,735} \code{4,768} \code{4,791} \code{4,821} \code{4,885} \code{4,887} \code{40} \code{40,659} \code{40,787} \code{401} \code{402} \code{403} \code{405} \code{406} \code{407} \code{408} \code{41} \code{41,207} \code{41,632} \code{410} \code{411} \code{412} \code{414} \code{415} \code{416,538} \code{417} \code{419} \code{42} \code{42,547} \code{420} \code{421} \code{422} \code{423} \code{424} \code{425} \code{426} \code{427} \code{429} \code{43} \code{43,658} \code{43,911} \code{430} \code{432} \code{432,839} \code{434} \code{435} \code{44} \code{44,045} \code{44,446} \code{44,741} \code{440} \code{441} \code{442} \code{447} \code{45} \code{45,191} \code{45,864} \code{45,965} \code{452} \code{454} \code{455} \code{456} \code{457} \code{46} \code{46,139} \code{46,999} \code{461} \code{464} \code{469} \code{47} \code{471} \code{472} \code{474} \code{476} \code{48} \code{48,208} \code{48,904} \code{480} \code{481} \code{485} \code{485,669} \code{486} \code{489} \code{49} \code{490,200} \code{490,589} \code{492} \code{493} \code{495} \code{499} \code{5} \code{5,074} \code{5,083} \code{5,100} \code{5,106} \code{5,121} \code{5,190} \code{5,211} \code{5,246} \code{5,268} \code{5,272} \code{5,306} \code{5,350} \code{5,466} \code{5,533} \code{5,657} \code{5,674} \code{5,736} \code{5,763} \code{5,829} \code{5,835} \code{5,844} \code{5,850} \code{5,870} \code{5,879} \code{5,948} \code{5,987} \code{50} \code{50,693} \code{50,817} \code{503} \code{505} \code{509} \code{51} \code{51,144} \code{51,193} \code{510} \code{512,975} \code{514} \code{515} \code{516} \code{518} \code{52} \code{522} \code{523} \code{525} \code{525,324} \code{526} \code{527} \code{529} \code{53} \code{53,154} \code{53,589} \code{530} \code{532} \code{533} \code{536} \code{537} \code{54} \code{54,174} \code{540,800} \code{546} \code{547} \code{548} \code{549} \code{55} \code{55,139} \code{55,183} \code{55,692} \code{55,910} \code{552} \code{552,022} \code{553} \code{555} \code{556} \code{56} \code{561} \code{564} \code{565} \code{567} \code{568} \code{57} \code{57,435} \code{571} \code{572} \code{575} \code{577} \code{579} \code{579,551} \code{58} \code{584} \code{585} \code{587} \code{59} \code{59,187} \code{59,566} \code{592} \code{594} \code{595} \code{597} \code{599} \code{6} \code{6,072} \code{6,082} \code{6,116} \code{6,143} \code{6,162} \code{6,180} \code{6,222} \code{6,242} \code{6,270} \code{6,308} \code{6,337} \code{6,380} \code{6,389} \code{6,394} \code{6,457} \code{6,474} \code{6,487} \code{6,488} \code{6,509} \code{6,514} \code{6,551} \code{6,604} \code{6,697} \code{6,747} \code{6,748} \code{6,757} \code{6,830} \code{6,850} \code{6,872} \code{6,891} \code{6,937} \code{6,969} \code{6,999} \code{60} \code{60,309} \code{601} \code{603} \code{608} \code{61} \code{61,036} \code{61,745} \code{610} \code{614} \code{615} \code{62} \code{620} \code{621} \code{623} \code{626} \code{627} \code{629} \code{63} \code{631} \code{633} \code{635} \code{636} \code{638} \code{639} \code{64} \code{64,356} \code{64,422} \code{641} \code{649} \code{65} \code{65,324} \code{650} \code{653} \code{659} \code{66} \code{661} \code{668} \code{669} \code{67} \code{67,486} \code{673} \code{679} \code{68} \code{68,182} \code{68,221} \code{68,603} \code{682} \code{684} \code{686} \code{689} \code{69} \code{690} \code{699} \code{7} \code{7,005} \code{7,008} \code{7,086} \code{7,150} \code{7,166} \code{7,194} \code{7,198} \code{7,332} \code{7,350} \code{7,400} \code{7,620} \code{7,667} \code{7,722} \code{7,861} \code{7,904} \code{7,921} \code{70} \code{702} \code{704} \code{705} \code{705,536} \code{71} \code{71,039} \code{710,330} \code{714} \code{717} \code{72} \code{72,359} \code{72,961} \code{724} \code{725} \code{73} \code{73,557} \code{73,949} \code{74} \code{74,001} \code{74,191} \code{74,645} \code{74,903} \code{746} \code{747} \code{748,051} \code{75} \code{757} \code{759} \code{76} \code{76,627} \code{762} \code{763} \code{77} \code{77,340} \code{774} \code{776,583} \code{778} \code{779} \code{78} \code{780} \code{781} \code{783} \code{785} \code{786} \code{787} \code{789} \code{79} \code{79,352} \code{791,616} \code{792} \code{794} \code{796} \code{799} \code{8} \code{8,025} \code{8,034} \code{8,084} \code{8,148} \code{8,257} \code{8,275} \code{8,280} \code{8,297} \code{8,328} \code{8,349} \code{8,386} \code{8,416} \code{8,481} \code{8,508} \code{8,540} \code{8,597} \code{8,742} \code{8,746} \code{8,789} \code{8,962} \code{8,963} \code{80} \code{80,991} \code{81} \code{81,399} \code{810} \code{814} \code{82} \code{82,744} \code{820,614} \code{822} \code{825} \code{826} \code{827} \code{83} \code{833} \code{836} \code{837} \code{84} \code{84,186} \code{84,493} \code{842} \code{846} \code{85} \code{85,643} \code{85,838} \code{852} \code{856} \code{857} \code{859} \code{86} \code{86,353} \code{86,571} \code{861} \code{867} \code{87} \code{877} \code{88} \code{881} \code{888} \code{89} \code{891} \code{896} \code{9} \code{9,060} \code{9,093} \code{9,248} \code{9,328} \code{9,447} \code{9,457} \code{9,532} \code{9,539} \code{9,645} \code{9,761} \code{9,867} \code{9,956} \code{90} \code{902} \code{91} \code{91,069} \code{91,384} \code{92} \code{924} \code{925} \code{928} \code{929} \code{93} \code{931} \code{937} \code{94} \code{94,090} \code{95} \code{954} \code{955} \code{96} \code{96,699} \code{960} \code{963} \code{964} \code{966} \code{969} \code{97} \code{978} \code{98} \code{987} \code{99} \code{c} \code{-5} \code{-6} \code{1,015} \code{1,017} \code{1,022} \code{1,024} \code{1,032} \code{1,043} \code{1,047} \code{1,052} \code{1,054} \code{1,056} \code{1,061} \code{1,082} \code{1,089} \code{1,090} \code{1,093} \code{1,094} \code{1,105} \code{1,106} \code{1,109} \code{1,113} \code{1,121} \code{1,127} \code{1,136} \code{1,159} \code{1,165} \code{1,174} \code{1,178} \code{1,202} \code{1,210} \code{1,219} \code{1,220} \code{1,228} \code{1,235} \code{1,237} \code{1,242} \code{1,248} \code{1,249} \code{1,254} \code{1,258} \code{1,261} \code{1,271} \code{1,275} \code{1,276} \code{1,279} \code{1,291} \code{1,292} \code{1,301} \code{1,302} \code{1,326} \code{1,353} \code{1,354} \code{1,355} \code{1,360,380} \code{1,360,647} \code{1,362} \code{1,366} \code{1,368,810} \code{1,375} \code{1,378} \code{1,391} \code{1,394,520} \code{1,396} \code{1,399} \code{1,427} \code{1,429,101} \code{1,431} \code{1,438} \code{1,458} \code{1,459} \code{1,460} \code{1,474} \code{1,484} \code{1,490} \code{1,492} \code{1,503} \code{1,517} \code{1,523} \code{1,527} \code{1,535} \code{1,545} \code{1,546} \code{1,558} \code{1,562} \code{1,575} \code{1,589} \code{1,595} \code{1,597} \code{1,604} \code{1,608} \code{1,613} \code{1,623} \code{1,627} \code{1,641} \code{1,678} \code{1,680} \code{1,681} \code{1,699} \code{1,702} \code{1,705} \code{1,722} \code{1,757} \code{1,770} \code{1,785} \code{1,808} \code{1,821} \code{1,827} \code{1,830} \code{1,839} \code{1,855} \code{1,863} \code{1,875} \code{1,876} \code{1,880} \code{1,890} \code{1,896} \code{1,898} \code{1,899} \code{1,900} \code{1,902} \code{1,915} \code{1,922} \code{1,926} \code{1,927} \code{1,932} \code{1,947} \code{1,948} \code{1,952} \code{1,953} \code{1,977} \code{1,978} \code{1,983} \code{1,984} \code{1,988} \code{1,990} \code{1,996} \code{10,115} \code{10,225} \code{10,258} \code{10,453} \code{10,807} \code{10,913} \code{10,989} \code{103,211} \code{104,411} \code{106,095} \code{106,134} \code{11,094} \code{11,228} \code{11,285} \code{11,309} \code{11,374} \code{11,435} \code{11,492} \code{11,578} \code{11,597} \code{11,656} \code{11,729} \code{11,733} \code{11,774} \code{115,441} \code{116,055} \code{119,196} \code{12,054} \code{12,168} \code{12,358} \code{12,359} \code{12,386} \code{12,416} \code{12,510} \code{12,591} \code{12,834} \code{12,880} \code{12,885} \code{121} \code{121,007} \code{123,368} \code{124,146} \code{124,198} \code{127,097} \code{127,297} \code{13,002} \code{13,045} \code{13,122} \code{13,301} \code{13,361} \code{13,403} \code{13,519} \code{13,578} \code{13,615} \code{13,638} \code{13,645} \code{13,689} \code{13,928} \code{130,954} \code{134,989} \code{137} \code{137,976} \code{138,711} \code{138,989} \code{139} \code{14,147,693} \code{14,207} \code{14,243} \code{14,457} \code{14,565} \code{14,621} \code{14,663} \code{14,792} \code{14,838} \code{14,846} \code{14,895} \code{142,260} \code{142,618} \code{142,940} \code{146,750} \code{147,061} \code{148,028} \code{148,440} \code{15,461} \code{15,733} \code{15,827} \code{152,485} \code{153,003} \code{16,008} \code{16,039} \code{16,065} \code{16,168} \code{16,268} \code{16,270} \code{16,401} \code{16,410} \code{16,417} \code{16,442} \code{16,451} \code{16,569} \code{16,644} \code{16,723} \code{16,732} \code{16,972} \code{17,008} \code{17,128} \code{17,317} \code{17,459} \code{17,522} \code{17,540} \code{17,605} \code{17,958} \code{171,262} \code{171,421} \code{171,713} \code{172} \code{175,166} \code{176,180} \code{179,420} \code{18,023} \code{18,046} \code{18,143} \code{18,187} \code{18,803} \code{180,988} \code{181} \code{187,671} \code{188} \code{19,211} \code{19,227} \code{19,228} \code{19,346} \code{19,458} \code{19,569} \code{19,885} \code{194,476} \code{195} \code{195,439} \code{197} \code{2,008} \code{2,019} \code{2,032} \code{2,047} \code{2,049} \code{2,059} \code{2,062} \code{2,063} \code{2,073} \code{2,076} \code{2,080} \code{2,083} \code{2,086} \code{2,127} \code{2,152} \code{2,172} \code{2,174} \code{2,192} \code{2,197} \code{2,207} \code{2,209} \code{2,216} \code{2,233} \code{2,246,034} \code{2,248} \code{2,261} \code{2,273} \code{2,285} \code{2,301} \code{2,306} \code{2,312} \code{2,321} \code{2,325} \code{2,330} \code{2,333} \code{2,344} \code{2,350} \code{2,353} \code{2,359} \code{2,373} \code{2,381} \code{2,409} \code{2,419} \code{2,420} \code{2,429} \code{2,430} \code{2,467} \code{2,486} \code{2,505} \code{2,508} \code{2,514} \code{2,524} \code{2,527} \code{2,534} \code{2,537} \code{2,543} \code{2,551} \code{2,558} \code{2,580} \code{2,584} \code{2,592} \code{2,606} \code{2,617} \code{2,637} \code{2,654} \code{2,661} \code{2,664} \code{2,665} \code{2,674} \code{2,683} \code{2,686} \code{2,699} \code{2,701} \code{2,708} \code{2,711} \code{2,715} \code{2,729} \code{2,759} \code{2,799} \code{2,805} \code{2,812} \code{2,836} \code{2,844} \code{2,846} \code{2,859} \code{2,866} \code{2,922} \code{2,935} \code{2,939} \code{2,942} \code{2,987} \code{2,989} \code{2,993} \code{20,109} \code{20,420} \code{20,474} \code{20,728} \code{20,873} \code{20,900} \code{20,996} \code{200} \code{201,777} \code{21,579} \code{21,631} \code{21,694} \code{21,807} \code{21,817} \code{216} \code{22,190} \code{22,218} \code{22,435} \code{22,628} \code{22,826} \code{22,832} \code{226} \code{227} \code{229} \code{23,007} \code{23,032} \code{23,166} \code{23,290} \code{23,490} \code{23,657} \code{23,939} \code{23,944} \code{230,841} \code{231,526} \code{237} \code{239,129} \code{24,115} \code{24,419} \code{24,447} \code{24,566} \code{24,865} \code{240,475} \code{244,068} \code{25,273} \code{25,282} \code{25,416} \code{25,700} \code{25,815} \code{25,845} \code{255} \code{256} \code{26,095} \code{26,286} \code{26,682} \code{26,920} \code{269,823} \code{27,016} \code{28,463} \code{28,845} \code{282} \code{284} \code{29,762} \code{294} \code{297} \code{3,006} \code{3,053} \code{3,073} \code{3,074} \code{3,075} \code{3,080} \code{3,103} \code{3,122} \code{3,144} \code{3,173} \code{3,182} \code{3,226} \code{3,234} \code{3,236} \code{3,239} \code{3,241} \code{3,244} \code{3,255} \code{3,260} \code{3,263} \code{3,270} \code{3,297,765} \code{3,303} \code{3,310} \code{3,346} \code{3,369} \code{3,383} \code{3,436} \code{3,441} \code{3,449} \code{3,466} \code{3,476} \code{3,482} \code{3,500} \code{3,553} \code{3,596} \code{3,615} \code{3,625} \code{3,641} \code{3,688} \code{3,691} \code{3,734} \code{3,797} \code{3,806} \code{3,811} \code{3,844} \code{3,861} \code{3,866} \code{3,871} \code{3,892} \code{3,930} \code{3,981} \code{3,991} \code{30,068} \code{30,929} \code{302} \code{305} \code{31,353} \code{31,567} \code{311} \code{316} \code{318,766} \code{32,925} \code{323} \code{33,031} \code{33,071} \code{33,261} \code{33,372} \code{33,378} \code{33,832} \code{331} \code{333,183} \code{336,609} \code{337} \code{337,847} \code{34,582} \code{34,859} \code{340} \code{344} \code{346} \code{347} \code{349} \code{35,450} \code{35,685} \code{359,464} \code{36,281} \code{36,419} \code{36,574} \code{36,931} \code{360} \code{363} \code{37,020} \code{37,570} \code{37,651} \code{373} \code{38,379} \code{38,432} \code{38,469} \code{38,865} \code{380} \code{381} \code{383} \code{39,138} \code{39,573} \code{39,985} \code{394} \code{4,012} \code{4,074} \code{4,086} \code{4,111} \code{4,151} \code{4,164} \code{4,179} \code{4,203} \code{4,277} \code{4,282} \code{4,312} \code{4,334} \code{4,356} \code{4,363} \code{4,368} \code{4,374} \code{4,383} \code{4,394} \code{4,395} \code{4,404} \code{4,445} \code{4,447} \code{4,490} \code{4,542} \code{4,555} \code{4,613} \code{4,685} \code{4,695} \code{4,699} \code{4,761} \code{4,775} \code{4,777} \code{4,941} \code{4,958} \code{4,963} \code{4,976} \code{4,998} \code{40,256} \code{404} \code{409} \code{41,277} \code{41,656} \code{41,760} \code{41,959} \code{418} \code{42,273} \code{428} \code{43,128} \code{43,594} \code{431} \code{438} \code{439} \code{443} \code{445} \code{446} \code{449} \code{45,155} \code{45,214} \code{451} \code{453} \code{46,872} \code{460} \code{463} \code{466} \code{467} \code{468} \code{47,544} \code{47,600} \code{477} \code{479} \code{48,786} \code{482} \code{483} \code{484} \code{487} \code{49,029} \code{49,758} \code{49,892} \code{490} \code{491} \code{494,639} \code{499,048} \code{5,004} \code{5,032} \code{5,089} \code{5,124} \code{5,158} \code{5,159} \code{5,227} \code{5,270} \code{5,286} \code{5,302} \code{5,330} \code{5,384} \code{5,385} \code{5,397} \code{5,469} \code{5,477} \code{5,556} \code{5,588} \code{5,604} \code{5,625} \code{5,697} \code{5,704} \code{5,719} \code{5,750} \code{5,761} \code{5,796} \code{5,800} \code{5,864} \code{5,884} \code{5,913} \code{5,915} \code{5,991} \code{50,649} \code{50,736} \code{500} \code{501} \code{502} \code{504} \code{506} \code{507} \code{508} \code{51,513} \code{51,814} \code{512,498} \code{52,303} \code{52,751} \code{524} \code{53,120} \code{53,536} \code{53,984} \code{530,383} \code{535,257} \code{54,150} \code{541} \code{542} \code{543} \code{55,194} \code{55,799} \code{557} \code{56,461} \code{562} \code{563} \code{566} \code{569} \code{570} \code{570,003} \code{573,967} \code{576} \code{578} \code{58,616} \code{581} \code{589} \code{590} \code{596} \code{596,058} \code{6,010} \code{6,015} \code{6,097} \code{6,100} \code{6,102} \code{6,176} \code{6,181} \code{6,202} \code{6,206} \code{6,299} \code{6,306} \code{6,328} \code{6,347} \code{6,355} \code{6,370} \code{6,374} \code{6,452} \code{6,476} \code{6,483} \code{6,505} \code{6,649} \code{6,679} \code{6,724} \code{6,782} \code{6,841} \code{6,862} \code{6,864} \code{6,965} \code{6,973} \code{604} \code{605} \code{61,403} \code{61,513} \code{611} \code{616} \code{619} \code{62,041} \code{62,096} \code{62,918} \code{624} \code{628} \code{63,305} \code{630} \code{632} \code{644} \code{645} \code{648} \code{649,527} \code{65,533} \code{655} \code{657} \code{66,383} \code{66,552} \code{660} \code{662} \code{664} \code{667} \code{67,165} \code{677} \code{680} \code{681} \code{685} \code{687} \code{688} \code{69,241} \code{69,450} \code{69,971} \code{693} \code{7,025} \code{7,026} \code{7,045} \code{7,068} \code{7,073} \code{7,103} \code{7,138} \code{7,139} \code{7,213} \code{7,263} \code{7,309} \code{7,321} \code{7,333} \code{7,375} \code{7,406} \code{7,420} \code{7,449} \code{7,513} \code{7,558} \code{7,586} \code{7,637} \code{7,745} \code{7,760} \code{7,787} \code{7,803} \code{7,937} \code{70,609} \code{700} \code{701} \code{709} \code{71,494} \code{715} \code{716} \code{718} \code{720} \code{721} \code{727} \code{73,186} \code{736} \code{738,323} \code{739} \code{750} \code{751} \code{76,020} \code{761} \code{764} \code{767} \code{768} \code{770} \code{773} \code{78,166} \code{782} \code{79,184} \code{79,331} \code{79,524} \code{8,013} \code{8,096} \code{8,107} \code{8,160} \code{8,170} \code{8,194} \code{8,263} \code{8,282} \code{8,473} \code{8,514} \code{8,531} \code{8,551} \code{8,637} \code{8,695} \code{8,720} \code{8,727} \code{8,754} \code{8,778} \code{8,877} \code{8,956} \code{8,975} \code{80,001} \code{80,135} \code{80,253} \code{802} \code{804} \code{809} \code{81,398} \code{815} \code{819} \code{82,631} \code{82,632} \code{820} \code{824} \code{825,517} \code{827,418} \code{828} \code{83,427} \code{831} \code{834} \code{845} \code{847,951} \code{85,754} \code{850} \code{854} \code{855} \code{858} \code{86,568} \code{862} \code{87,488} \code{87,524} \code{883} \code{885} \code{897} \code{897,832} \code{899} \code{9,050} \code{9,102} \code{9,255} \code{9,294} \code{9,314} \code{9,325} \code{9,435} \code{9,473} \code{9,511} \code{9,568} \code{9,594} \code{9,807} \code{9,897} \code{9,951} \code{9,972} \code{90,453} \code{901} \code{906} \code{907} \code{91,051} \code{911} \code{913} \code{916} \code{923,399} \code{927} \code{93,931} \code{931,876} \code{932} \code{935} \code{936} \code{939} \code{94,836} \code{94,846} \code{941} \code{943} \code{952} \code{959} \code{962} \code{97,036} \code{98,232} \code{981} \code{982} \code{991} \code{997} \code{999} \code{-1} \code{-17} \code{-23} \code{1,006,275} \code{1,014} \code{1,016} \code{1,021} \code{1,034} \code{1,035} \code{1,036} \code{1,049} \code{1,055} \code{1,062} \code{1,065} \code{1,067} \code{1,075} \code{1,078} \code{1,080} \code{1,092} \code{1,104} \code{1,112} \code{1,115} \code{1,117} \code{1,117,614} \code{1,118} \code{1,129} \code{1,132} \code{1,133} \code{1,142} \code{1,150} \code{1,157} \code{1,158} \code{1,160} \code{1,168} \code{1,170} \code{1,184} \code{1,204} \code{1,205,127} \code{1,209} \code{1,216} \code{1,229} \code{1,256} \code{1,280} \code{1,293} \code{1,311} \code{1,313,815} \code{1,324} \code{1,333,204} \code{1,343} \code{1,344} \code{1,350} \code{1,356} \code{1,369,577} \code{1,370} \code{1,373} \code{1,379} \code{1,381} \code{1,393} \code{1,395} \code{1,402} \code{1,410} \code{1,424} \code{1,439} \code{1,443} \code{1,449} \code{1,453} \code{1,457} \code{1,475} \code{1,477} \code{1,483} \code{1,485} \code{1,488} \code{1,493} \code{1,497} \code{1,501} \code{1,508} \code{1,513} \code{1,519} \code{1,521} \code{1,531} \code{1,550} \code{1,554} \code{1,570} \code{1,578} \code{1,579} \code{1,581} \code{1,584} \code{1,590} \code{1,596} \code{1,617} \code{1,620} \code{1,631} \code{1,635} \code{1,639} \code{1,645} \code{1,659} \code{1,670,204} \code{1,684} \code{1,690} \code{1,691} \code{1,713} \code{1,717} \code{1,721} \code{1,721,307} \code{1,740} \code{1,742} \code{1,747} \code{1,754} \code{1,764} \code{1,768} \code{1,780} \code{1,788} \code{1,792} \code{1,793} \code{1,810} \code{1,817} \code{1,843} \code{1,844,047} \code{1,849,783} \code{1,850} \code{1,851} \code{1,855,404} \code{1,865} \code{1,881} \code{1,883} \code{1,886} \code{1,891} \code{1,910} \code{1,931} \code{1,944} \code{1,956} \code{1,961} \code{1,972} \code{1,985} \code{1,992} \code{1,997} \code{10,000} \code{10,040} \code{10,166} \code{10,185} \code{10,296} \code{10,312} \code{10,319} \code{10,344} \code{10,358} \code{10,372} \code{10,443} \code{10,456} \code{10,558} \code{10,643} \code{10,690} \code{10,710} \code{10,773} \code{10,784} \code{10,793} \code{10,808} \code{10,832} \code{10,903} \code{10,958} \code{106,007} \code{106,191} \code{106,984} \code{107,018} \code{107,395} \code{107,412} \code{108,277} \code{109,018} \code{109,190} \code{109,505} \code{109,876} \code{11,149} \code{11,205} \code{11,373} \code{11,378} \code{11,480} \code{11,595} \code{11,683} \code{11,711} \code{11,742} \code{11,786} \code{11,834} \code{11,963} \code{11,980} \code{113,694} \code{116,295} \code{117,333} \code{118,673} \code{119,125} \code{119,172} \code{119,715} \code{12,003} \code{12,038} \code{12,096} \code{12,153} \code{12,225} \code{12,226} \code{12,366} \code{12,491} \code{12,492} \code{12,772} \code{12,828} \code{12,886} \code{12,909} \code{12,966} \code{12,985} \code{122,169} \code{123,220} \code{124,972} \code{126,178} \code{127,034} \code{128,139} \code{128,375} \code{128,483} \code{13,082} \code{13,115} \code{13,163} \code{13,181} \code{13,453} \code{13,587} \code{13,656} \code{13,738} \code{13,810} \code{13,812} \code{13,817} \code{133,346} \code{14,296} \code{14,359} \code{14,530} \code{14,650} \code{14,812} \code{142,948} \code{143,312} \code{15,296} \code{15,463} \code{15,475} \code{15,699} \code{15,750} \code{15,913} \code{152,900} \code{154,751} \code{159,575} \code{159,934} \code{16,032} \code{16,081} \code{16,148} \code{16,198} \code{16,253} \code{16,438} \code{16,783} \code{16,905} \code{16,987} \code{167,894} \code{17,155} \code{17,343} \code{17,365} \code{17,420} \code{17,564} \code{17,625} \code{17,761} \code{17,849} \code{17,985} \code{170,667} \code{171,140} \code{172,170} \code{178,304} \code{18,150} \code{18,218} \code{18,263} \code{18,433} \code{18,515} \code{18,640} \code{18,644} \code{18,648} \code{18,698} \code{18,699} \code{18,748} \code{18,845} \code{181,419} \code{182,193} \code{183,425} \code{184,358} \code{184,871} \code{186,611} \code{188,203} \code{189,594} \code{19,033} \code{19,052} \code{19,090} \code{19,163} \code{19,217,763} \code{19,325} \code{19,676} \code{19,702} \code{19,816} \code{19,914} \code{190} \code{194,939} \code{2,000} \code{2,007} \code{2,009} \code{2,013} \code{2,021} \code{2,035} \code{2,043} \code{2,056} \code{2,070} \code{2,090} \code{2,093} \code{2,099} \code{2,107} \code{2,117} \code{2,120} \code{2,124} \code{2,129} \code{2,131} \code{2,138} \code{2,154} \code{2,175} \code{2,177} \code{2,179} \code{2,180} \code{2,191} \code{2,199} \code{2,214} \code{2,223} \code{2,238} \code{2,251} \code{2,263} \code{2,271} \code{2,272} \code{2,282} \code{2,299} \code{2,308} \code{2,326} \code{2,332} \code{2,335} \code{2,347} \code{2,352} \code{2,354} \code{2,358} \code{2,367} \code{2,385} \code{2,395} \code{2,398} \code{2,412} \code{2,422} \code{2,427} \code{2,432} \code{2,476} \code{2,512} \code{2,513} \code{2,518} \code{2,523} \code{2,529} \code{2,530} \code{2,535} \code{2,539} \code{2,556} \code{2,567} \code{2,613} \code{2,627} \code{2,638} \code{2,671} \code{2,675} \code{2,694} \code{2,707} \code{2,710} \code{2,713} \code{2,735} \code{2,751} \code{2,752} \code{2,756} \code{2,764} \code{2,815} \code{2,834} \code{2,875} \code{2,888} \code{2,889} \code{2,905} \code{2,912} \code{2,921} \code{2,940} \code{2,965} \code{2,968} \code{20,396} \code{20,488} \code{20,492} \code{20,496} \code{20,588} \code{205,569} \code{206,733} \code{207,035} \code{208,490} \code{21,243} \code{21,335} \code{21,348} \code{21,446} \code{21,555} \code{21,594} \code{21,598} \code{21,760} \code{21,840} \code{212,775} \code{213,956} \code{214,612} \code{22,141} \code{22,253} \code{22,555} \code{221,507} \code{228,604} \code{23,383} \code{23,495} \code{23,518} \code{23,551} \code{23,748} \code{231,493} \code{232,664} \code{234,295} \code{24,022} \code{24,207} \code{24,445} \code{24,678} \code{24,737} \code{24,811} \code{24,852} \code{244} \code{245,985} \code{249,435} \code{25,001} \code{25,717} \code{25,764} \code{25,976} \code{253} \code{253,978} \code{26,149} \code{26,575} \code{26,612} \code{26,654} \code{26,701} \code{26,877} \code{27,230} \code{27,274} \code{27,661} \code{272} \code{28,281} \code{28,284} \code{28,582} \code{28,664} \code{28,824} \code{28,835} \code{29,004} \code{29,547} \code{29,882} \code{291,850} \code{294,833} \code{3,000} \code{3,013} \code{3,058} \code{3,091} \code{3,102} \code{3,107} \code{3,120} \code{3,134} \code{3,134,244} \code{3,135} \code{3,141} \code{3,143} \code{3,149} \code{3,165} \code{3,168} \code{3,211} \code{3,219} \code{3,221} \code{3,231} \code{3,242} \code{3,296} \code{3,308} \code{3,315} \code{3,320} \code{3,365} \code{3,375} \code{3,380} \code{3,393} \code{3,408} \code{3,416} \code{3,417} \code{3,455} \code{3,481} \code{3,487} \code{3,494} \code{3,519} \code{3,526} \code{3,554} \code{3,567} \code{3,568} \code{3,579} \code{3,599} \code{3,621} \code{3,622} \code{3,626} \code{3,647} \code{3,650} \code{3,664} \code{3,681} \code{3,687} \code{3,700} \code{3,701} \code{3,767} \code{3,770} \code{3,790} \code{3,793} \code{3,817} \code{3,822} \code{3,852} \code{3,864} \code{3,878} \code{3,914} \code{3,953} \code{3,972} \code{3,999} \code{30,122} \code{30,175} \code{30,234} \code{30,281} \code{30,886} \code{30,985} \code{300,978} \code{303} \code{311,349} \code{32,226} \code{32,472} \code{323,142} \code{33,226} \code{33,342} \code{33,470} \code{33,586} \code{33,599} \code{33,684} \code{33,713} \code{33,841} \code{33,859} \code{334,912} \code{339} \code{34,262} \code{34,484} \code{34,896} \code{34,966} \code{341} \code{35,098} \code{35,419} \code{35,770} \code{350,592} \code{36,013} \code{36,102} \code{36,563} \code{36,680} \code{365} \code{37,145} \code{37,661} \code{371} \code{378} \code{38,158} \code{38,495} \code{38,641} \code{38,693} \code{38,830} \code{381,921} \code{39,521} \code{392} \code{4,014} \code{4,024} \code{4,029} \code{4,041} \code{4,058} \code{4,073} \code{4,119} \code{4,125} \code{4,143} \code{4,153} \code{4,215,192} \code{4,228} \code{4,292} \code{4,314} \code{4,315} \code{4,328} \code{4,331} \code{4,376} \code{4,382} \code{4,424} \code{4,464} \code{4,471} \code{4,499} \code{4,514} \code{4,526} \code{4,566} \code{4,578} \code{4,590} \code{4,593} \code{4,606} \code{4,650} \code{4,658} \code{4,700} \code{4,710} \code{4,745} \code{4,746} \code{4,790} \code{4,812} \code{4,817} \code{4,850} \code{4,859} \code{4,920} \code{4,965} \code{4,979} \code{40,280} \code{40,752} \code{400} \code{41,091} \code{41,126} \code{41,964} \code{413} \code{414,802} \code{415,562} \code{416} \code{42,147} \code{42,254} \code{42,275} \code{424,395} \code{43,003} \code{43,298} \code{433} \code{436} \code{44,106} \code{45,257} \code{450} \code{455,082} \code{458} \code{46,670} \code{462} \code{465} \code{47,327} \code{470} \code{473} \code{478} \code{48,142} \code{48,574} \code{48,777} \code{48,779} \code{481,570} \code{488} \code{49,587} \code{496} \code{498} \code{5,005} \code{5,024} \code{5,046} \code{5,057} \code{5,061} \code{5,071} \code{5,072} \code{5,079} \code{5,091} \code{5,123} \code{5,265} \code{5,276} \code{5,283} \code{5,285} \code{5,289} \code{5,321} \code{5,328} \code{5,354} \code{5,357} \code{5,370} \code{5,409} \code{5,483} \code{5,491} \code{5,505} \code{5,540} \code{5,559} \code{5,613} \code{5,632} \code{5,644} \code{5,716} \code{5,718} \code{5,737} \code{5,793} \code{5,798} \code{5,817} \code{5,841} \code{5,851} \code{5,852} \code{5,865} \code{5,889} \code{5,897} \code{5,906} \code{5,920} \code{5,935} \code{5,939} \code{5,970} \code{50,229} \code{50,489} \code{51,007} \code{51,547} \code{51,692} \code{511} \code{512} \code{517} \code{52,716} \code{52,991} \code{520} \code{521} \code{528} \code{53,429} \code{53,843} \code{534} \code{535} \code{538} \code{54,395} \code{54,916} \code{540} \code{545} \code{55,286} \code{55,511} \code{55,999} \code{550} \code{551} \code{554} \code{559} \code{56,011} \code{56,240} \code{56,441} \code{57,442} \code{574} \code{58,380} \code{59,149} \code{59,488} \code{591} \code{593} \code{6,008} \code{6,009} \code{6,043} \code{6,057} \code{6,163} \code{6,217} \code{6,248} \code{6,263} \code{6,284} \code{6,334} \code{6,352} \code{6,398} \code{6,429} \code{6,479} \code{6,498} \code{6,513} \code{6,519} \code{6,522} \code{6,555} \code{6,608} \code{6,639} \code{6,650} \code{6,696} \code{6,743} \code{6,751} \code{6,803} \code{6,881} \code{6,915} \code{6,935} \code{6,955} \code{60,897} \code{60,932} \code{600} \code{606} \code{607} \code{609} \code{61,329} \code{612} \code{618} \code{62,366} \code{62,841} \code{62,886} \code{620,208} \code{625} \code{64,288} \code{642} \code{646} \code{65,621} \code{652} \code{658} \code{66,706} \code{66,931} \code{663} \code{663,120} \code{666} \code{672} \code{672,614} \code{675} \code{676} \code{678} \code{68,512} \code{68,602} \code{691} \code{695} \code{698} \code{7,125} \code{7,134} \code{7,137} \code{7,144} \code{7,181} \code{7,205} \code{7,249} \code{7,258} \code{7,260} \code{7,261} \code{7,262} \code{7,292} \code{7,329} \code{7,339} \code{7,427} \code{7,428} \code{7,565} \code{7,635} \code{7,708} \code{7,734} \code{7,770} \code{7,819} \code{7,893} \code{7,895} \code{7,925} \code{7,953} \code{7,961} \code{703} \code{705,596} \code{712} \code{726} \code{728} \code{729} \code{73,902} \code{733} \code{734} \code{737} \code{738,029} \code{749} \code{758} \code{76,524} \code{760} \code{765} \code{766} \code{769} \code{77,394} \code{78,382} \code{79,207} \code{791} \code{791,064} \code{8,002} \code{8,012} \code{8,065} \code{8,132} \code{8,165} \code{8,218} \code{8,261} \code{8,271} \code{8,306} \code{8,363} \code{8,372} \code{8,397} \code{8,438} \code{8,504} \code{8,533} \code{8,583} \code{8,596} \code{8,671} \code{8,737} \code{8,764} \code{8,806} \code{8,858} \code{8,879} \code{8,912} \code{8,915} \code{8,959} \code{8,976} \code{80,232} \code{800} \code{81,638} \code{811} \code{812} \code{813} \code{82,203} \code{82,965} \code{82,969} \code{821} \code{832} \code{834,912} \code{841} \code{843} \code{85,644} \code{851} \code{86,349} \code{860} \code{861,265} \code{87,344} \code{870,535} \code{878} \code{886} \code{89,726} \code{9,006} \code{9,139} \code{9,263} \code{9,296} \code{9,305} \code{9,331} \code{9,346} \code{9,375} \code{9,581} \code{9,591} \code{9,599} \code{9,606} \code{9,697} \code{9,823} \code{9,838} \code{9,941} \code{9,955} \code{90,335} \code{90,502} \code{904} \code{905} \code{918} \code{920} \code{94,843} \code{942} \code{944} \code{945} \code{946} \code{95,825} \code{96,970} \code{968} \code{97,290} \code{98,845} \code{980} \code{988} \code{988,526} \code{989} \code{99,549} \code{99,910} \code{998} \code{-18} \code{-3} \code{1,000} \code{1,004} \code{1,009} \code{1,020} \code{1,029} \code{1,033} \code{1,037} \code{1,040} \code{1,042} \code{1,053} \code{1,066} \code{1,068} \code{1,071} \code{1,072} \code{1,072,312} \code{1,079} \code{1,098} \code{1,104,430} \code{1,114,549} \code{1,137} \code{1,139} \code{1,148} \code{1,162} \code{1,164,799} \code{1,166} \code{1,173} \code{1,175} \code{1,190} \code{1,192} \code{1,207} \code{1,211} \code{1,217,861} \code{1,225} \code{1,247} \code{1,250} \code{1,281} \code{1,297,152} \code{1,303} \code{1,327} \code{1,329} \code{1,334} \code{1,338} \code{1,345} \code{1,348} \code{1,382} \code{1,392} \code{1,394} \code{1,412} \code{1,413} \code{1,414} \code{1,417} \code{1,428} \code{1,454} \code{1,472} \code{1,489} \code{1,494} \code{1,510} \code{1,515,477} \code{1,532} \code{1,538} \code{1,552} \code{1,553} \code{1,556} \code{1,559} \code{1,570,955} \code{1,572} \code{1,580} \code{1,585} \code{1,588} \code{1,601} \code{1,609} \code{1,616,314} \code{1,625} \code{1,633} \code{1,638} \code{1,648} \code{1,654} \code{1,655} \code{1,658} \code{1,675} \code{1,694} \code{1,703} \code{1,714} \code{1,724} \code{1,725} \code{1,738} \code{1,745} \code{1,750,706} \code{1,766} \code{1,769} \code{1,771} \code{1,777} \code{1,779} \code{1,787} \code{1,790} \code{1,812} \code{1,815} \code{1,835} \code{1,840} \code{1,849} \code{1,853} \code{1,870} \code{1,893} \code{1,921} \code{1,925} \code{1,929} \code{1,930} \code{1,942} \code{1,967} \code{1,982} \code{1,999} \code{10,220} \code{10,454} \code{10,492} \code{10,498} \code{10,524} \code{10,631} \code{10,703} \code{10,760} \code{10,873} \code{10,874} \code{10,887} \code{100,111} \code{100,676} \code{103,412} \code{109,380} \code{11,008} \code{11,055} \code{11,058} \code{11,096} \code{11,141} \code{11,171} \code{11,220} \code{11,354} \code{11,420} \code{11,433} \code{11,465} \code{11,570} \code{11,598} \code{11,604} \code{11,655} \code{11,700} \code{11,798} \code{11,803} \code{11,816} \code{11,828} \code{11,910} \code{110,276} \code{110,364} \code{115,549} \code{117,532} \code{118,093} \code{118,616} \code{12,034} \code{12,071} \code{12,154} \code{12,198} \code{12,211} \code{12,323} \code{12,352} \code{12,459} \code{12,462} \code{12,487} \code{12,504} \code{12,636} \code{12,706} \code{12,723} \code{12,733} \code{12,754} \code{12,764} \code{120,253} \code{120,514} \code{124,121} \code{124,822} \code{124,885} \code{124,949} \code{125,062} \code{128,637} \code{13,036} \code{13,049} \code{13,236} \code{13,245} \code{13,349} \code{13,426} \code{13,469} \code{13,471} \code{13,715} \code{13,752} \code{13,882} \code{13,922} \code{13,948} \code{13,991} \code{130,302} \code{133,135} \code{133,252} \code{133,857} \code{134,270} \code{138,043} \code{139,355} \code{139,644} \code{14,097} \code{14,149} \code{14,305} \code{14,610} \code{14,719} \code{14,741} \code{14,756} \code{14,858} \code{14,867} \code{14,878} \code{14,891} \code{14,970} \code{142,013} \code{145,939} \code{149,628} \code{15,019} \code{15,198} \code{15,238} \code{15,249} \code{15,299} \code{15,477} \code{15,630} \code{15,631} \code{15,694} \code{15,874} \code{15,877} \code{15,964} \code{153,187} \code{153,198} \code{153,706} \code{154,493} \code{154,957} \code{154,959} \code{155,680} \code{157,632} \code{16,405} \code{16,735} \code{16,829} \code{16,898} \code{16,927} \code{164,074} \code{167,036} \code{169,831} \code{17,058} \code{17,094} \code{17,103} \code{17,186} \code{17,455} \code{17,555} \code{17,689} \code{17,726} \code{17,919} \code{174,048} \code{179,350} \code{18,059} \code{18,307} \code{18,375} \code{18,629} \code{18,666} \code{18,689} \code{18,743} \code{18,867} \code{18,927} \code{18,936} \code{19,286} \code{19,383} \code{19,443} \code{19,625} \code{19,783} \code{19,817} \code{19,973} \code{193,573} \code{195,241} \code{195,677} \code{197,106} \code{197,487} \code{2,009,672} \code{2,014} \code{2,015} \code{2,033} \code{2,045} \code{2,060} \code{2,061} \code{2,079} \code{2,084} \code{2,085} \code{2,092} \code{2,109,701} \code{2,126,689} \code{2,136} \code{2,137} \code{2,145,013} \code{2,156} \code{2,164} \code{2,173} \code{2,193} \code{2,222} \code{2,226} \code{2,241} \code{2,243} \code{2,244} \code{2,257} \code{2,262} \code{2,264} \code{2,265,854} \code{2,284} \code{2,298} \code{2,313} \code{2,322} \code{2,324} \code{2,327} \code{2,342} \code{2,382} \code{2,397} \code{2,416} \code{2,439} \code{2,447} \code{2,453} \code{2,464} \code{2,465} \code{2,468} \code{2,472} \code{2,488} \code{2,493} \code{2,504} \code{2,568} \code{2,574} \code{2,607} \code{2,621} \code{2,652} \code{2,656} \code{2,667} \code{2,668} \code{2,679} \code{2,717} \code{2,725} \code{2,732} \code{2,736} \code{2,746} \code{2,757} \code{2,771} \code{2,773} \code{2,777} \code{2,780} \code{2,784} \code{2,785} \code{2,786} \code{2,798} \code{2,819} \code{2,835} \code{2,839} \code{2,872} \code{2,874} \code{2,896} \code{2,899} \code{2,900} \code{2,909} \code{2,913} \code{2,931} \code{2,950} \code{2,971} \code{2,976} \code{2,982} \code{20,276} \code{20,339} \code{20,406} \code{20,670} \code{20,694} \code{200,677} \code{201,799} \code{202,361} \code{204,938} \code{206,145} \code{206,700} \code{21,258} \code{21,314} \code{21,325} \code{21,340} \code{21,519} \code{21,856} \code{211,023} \code{212,505} \code{214,950} \code{22,264} \code{22,307} \code{22,421} \code{22,547} \code{22,577} \code{22,600} \code{22,974} \code{225,009} \code{23,247} \code{23,484,559} \code{23,514} \code{23,525} \code{23,618} \code{23,652} \code{23,706} \code{23,882} \code{23,973} \code{230,414} \code{233,231} \code{233,574} \code{239} \code{24,037} \code{24,144} \code{24,147} \code{24,208} \code{24,303} \code{24,346} \code{24,390} \code{24,426} \code{24,448} \code{24,647} \code{24,982} \code{246,106} \code{25,210} \code{256,334} \code{257,737} \code{26,289} \code{26,687} \code{26,746} \code{26,842} \code{260,588} \code{262} \code{264,538} \code{268,262} \code{268,959} \code{27,546} \code{27,743} \code{272,848} \code{274,253} \code{274,739} \code{276} \code{279,210} \code{28,128} \code{28,351} \code{28,368} \code{284,266} \code{29,090} \code{29,091} \code{29,110} \code{29,446} \code{29,551} \code{3,028} \code{3,040} \code{3,047} \code{3,114} \code{3,153} \code{3,157} \code{3,159} \code{3,167} \code{3,174} \code{3,176} \code{3,179} \code{3,207} \code{3,212} \code{3,233} \code{3,237} \code{3,240} \code{3,248} \code{3,258} \code{3,274} \code{3,290} \code{3,306} \code{3,318} \code{3,329} \code{3,342} \code{3,357} \code{3,385} \code{3,386} \code{3,398} \code{3,423} \code{3,450} \code{3,451} \code{3,454} \code{3,457} \code{3,483} \code{3,523} \code{3,538} \code{3,563} \code{3,565} \code{3,629} \code{3,640} \code{3,663} \code{3,665} \code{3,694} \code{3,710} \code{3,721} \code{3,764,346} \code{3,778} \code{3,783} \code{3,802} \code{3,824} \code{3,826} \code{3,853} \code{3,857} \code{3,916} \code{3,935} \code{3,941} \code{3,951} \code{3,976} \code{3,978} \code{30,277} \code{30,379} \code{30,477} \code{307,344} \code{309,658} \code{31,062} \code{31,165} \code{31,371} \code{31,677} \code{31,949} \code{310,693} \code{32,527} \code{32,866} \code{324} \code{33,161} \code{33,344} \code{33,452} \code{34,153} \code{34,315} \code{34,369} \code{34,462} \code{34,651} \code{34,679} \code{34,856} \code{34,887} \code{344,669} \code{35,435} \code{35,691} \code{351} \code{36,227} \code{36,507} \code{36,743} \code{36,781} \code{360,178} \code{366} \code{367} \code{369,806} \code{37,350} \code{37,566} \code{37,922} \code{38,050} \code{38,076} \code{38,090} \code{38,150} \code{38,346} \code{38,616} \code{38,896} \code{380,933} \code{383,871} \code{39,030} \code{39,509} \code{399} \code{4,047} \code{4,126} \code{4,184} \code{4,185} \code{4,201} \code{4,204} \code{4,212} \code{4,238} \code{4,243} \code{4,249} \code{4,256} \code{4,266} \code{4,299} \code{4,309} \code{4,322} \code{4,325} \code{4,366} \code{4,367} \code{4,415} \code{4,436} \code{4,455} \code{4,465} \code{4,467} \code{4,475} \code{4,519} \code{4,560} \code{4,604} \code{4,646} \code{4,680} \code{4,689} \code{4,692} \code{4,723} \code{4,727} \code{4,731} \code{4,736} \code{4,778} \code{4,827} \code{4,832} \code{4,834} \code{4,857} \code{4,910,853} \code{4,916} \code{4,932} \code{4,955} \code{4,962} \code{400,889} \code{41,622} \code{41,716} \code{42,024} \code{42,041} \code{43,023} \code{43,062} \code{43,140} \code{43,248} \code{43,304} \code{43,319} \code{44,413} \code{444} \code{448} \code{45,596} \code{45,610} \code{454,478} \code{47,091} \code{47,099} \code{47,175} \code{47,279} \code{47,304} \code{47,355} \code{475} \code{48,345} \code{48,578} \code{49,199} \code{49,616} \code{49,901} \code{49,966} \code{494} \code{5,012} \code{5,045} \code{5,055} \code{5,076} \code{5,087} \code{5,104} \code{5,105} \code{5,111} \code{5,143} \code{5,194} \code{5,196} \code{5,231} \code{5,267} \code{5,274} \code{5,277} \code{5,290} \code{5,293} \code{5,352} \code{5,356} \code{5,379} \code{5,443} \code{5,525} \code{5,529} \code{5,538} \code{5,609} \code{5,610} \code{5,638} \code{5,690} \code{5,699} \code{5,708} \code{5,820} \code{5,849} \code{5,903} \code{5,911} \code{5,921} \code{5,941} \code{5,943} \code{5,957} \code{5,984} \code{50,431} \code{507,578} \code{513} \code{517,762} \code{519} \code{519,920} \code{52,283} \code{52,655} \code{52,666} \code{52,715} \code{521,430} \code{53,051} \code{53,658} \code{531} \code{534,075} \code{539} \code{54,272} \code{55,264} \code{55,324} \code{558} \code{560} \code{57,790} \code{57,986} \code{58,213} \code{58,378} \code{58,463} \code{58,836} \code{583} \code{586} \code{588} \code{59,651} \code{59,847} \code{6,063} \code{6,064} \code{6,067} \code{6,080} \code{6,084} \code{6,093} \code{6,113} \code{6,170} \code{6,185} \code{6,199} \code{6,216} \code{6,266} \code{6,341} \code{6,383} \code{6,386} \code{6,396} \code{6,397} \code{6,424} \code{6,570} \code{6,628} \code{6,646} \code{6,700} \code{6,712} \code{6,718} \code{6,793} \code{6,806} \code{6,859} \code{6,863} \code{6,902} \code{6,909} \code{6,921} \code{6,941} \code{6,987} \code{6,988} \code{60,252} \code{60,485} \code{602} \code{617} \code{62,398} \code{62,999} \code{63,564} \code{63,715} \code{637} \code{64,620} \code{643} \code{65,276} \code{65,397} \code{651} \code{654} \code{66,069} \code{66,171} \code{66,317} \code{67,702} \code{67,895} \code{670} \code{674} \code{68,021} \code{68,130} \code{68,591} \code{68,613} \code{68,905} \code{683} \code{683,229} \code{694} \code{694,382} \code{7,058} \code{7,090} \code{7,113} \code{7,130} \code{7,172} \code{7,211} \code{7,227} \code{7,231} \code{7,233} \code{7,299} \code{7,399} \code{7,457} \code{7,556} \code{7,571} \code{7,604} \code{7,625} \code{7,689} \code{7,697} \code{7,736} \code{7,764} \code{7,772} \code{7,788} \code{7,799} \code{7,829} \code{7,926} \code{7,940} \code{7,981} \code{706} \code{707} \code{71,275} \code{71,368} \code{71,396} \code{71,828} \code{71,863} \code{710} \code{713} \code{73,613} \code{737,754} \code{74,712} \code{74,930} \code{740} \code{743} \code{744} \code{745} \code{75,339} \code{756} \code{76,496} \code{76,527} \code{76,528} \code{76,758} \code{77,709} \code{77,710} \code{77,849} \code{772} \code{775,523} \code{78,801} \code{795} \code{8,224} \code{8,243} \code{8,259} \code{8,273} \code{8,290} \code{8,324} \code{8,335} \code{8,352} \code{8,364} \code{8,396} \code{8,398} \code{8,457} \code{8,468} \code{8,480} \code{8,552} \code{8,616} \code{8,623} \code{8,666} \code{8,744} \code{8,819} \code{8,823} \code{8,825} \code{8,856} \code{8,935} \code{8,978} \code{80,032} \code{803} \code{805} \code{806} \code{817} \code{823} \code{83,302} \code{83,303} \code{830} \code{835} \code{847} \code{849} \code{853} \code{86,537} \code{86,590} \code{869} \code{87,097} \code{87,450} \code{872} \code{880} \code{887} \code{89,187} \code{892} \code{894} \code{895} \code{898} \code{9,000} \code{9,016} \code{9,039} \code{9,070} \code{9,088} \code{9,246} \code{9,269} \code{9,336} \code{9,368} \code{9,396} \code{9,488} \code{9,512} \code{9,544} \code{9,569} \code{9,612} \code{9,616} \code{9,620} \code{9,622} \code{9,628} \code{9,660} \code{9,661} \code{9,695} \code{9,745} \code{9,809} \code{9,833} \code{9,852} \code{9,853} \code{9,878} \code{9,960} \code{90,281} \code{900} \code{909} \code{914} \code{914,014} \code{921} \code{922} \code{923} \code{930} \code{934} \code{934,790} \code{94,699} \code{951} \code{957,505} \code{965} \code{975} \code{98,654} \code{98,803} \code{98,807} \code{99,291} \code{990} \code{992} \code{-31} \code{-4} \code{-73} \code{1,003} \code{1,008} \code{1,023} \code{1,025} \code{1,027} \code{1,045} \code{1,058} \code{1,063} \code{1,064,478} \code{1,073} \code{1,074} \code{1,076} \code{1,086} \code{1,087} \code{1,101} \code{1,119} \code{1,122} \code{1,134} \code{1,161,463} \code{1,176} \code{1,177} \code{1,180} \code{1,182,212} \code{1,189} \code{1,193} \code{1,197} \code{1,197,069} \code{1,199} \code{1,217} \code{1,218} \code{1,232} \code{1,236} \code{1,244,227} \code{1,248,358} \code{1,262} \code{1,263} \code{1,266} \code{1,285} \code{1,295} \code{1,295,878} \code{1,308} \code{1,314,920} \code{1,317} \code{1,323} \code{1,335} \code{1,360} \code{1,369} \code{1,374} \code{1,388} \code{1,404} \code{1,440} \code{1,448} \code{1,464} \code{1,466} \code{1,476} \code{1,479} \code{1,506} \code{1,511} \code{1,512} \code{1,516} \code{1,522} \code{1,533} \code{1,552,803} \code{1,563} \code{1,574} \code{1,611} \code{1,612} \code{1,616} \code{1,621,245} \code{1,634} \code{1,644} \code{1,676} \code{1,682} \code{1,700} \code{1,712} \code{1,741} \code{1,760} \code{1,772} \code{1,776} \code{1,818} \code{1,822} \code{1,823} \code{1,840,921} \code{1,873,259} \code{1,877} \code{1,889} \code{1,895} \code{1,918} \code{1,939} \code{1,966} \code{1,969} \code{1,976} \code{1,991} \code{1,995} \code{10,016} \code{10,028} \code{10,029} \code{10,066} \code{10,125} \code{10,180} \code{10,210} \code{10,240} \code{10,377} \code{10,408} \code{10,458} \code{10,521} \code{10,537} \code{10,554} \code{10,639} \code{10,683} \code{10,728} \code{10,815} \code{10,836} \code{10,850} \code{10,879} \code{10,886} \code{10,926} \code{10,944} \code{10,962} \code{10,968} \code{100,423} \code{100,693} \code{100,979} \code{101,681} \code{103,075} \code{103,664} \code{104,740} \code{106,398} \code{108,309} \code{109,482} \code{11,116} \code{11,136} \code{11,195} \code{11,253} \code{11,265} \code{11,282} \code{11,340} \code{11,361} \code{11,379} \code{11,388} \code{11,546} \code{11,556} \code{11,716} \code{11,719} \code{11,745} \code{11,799} \code{11,861} \code{11,935} \code{11,956} \code{11,973} \code{111,555} \code{113,991} \code{114,015} \code{116,213} \code{116,791} \code{117,210} \code{117,504} \code{118,500} \code{118,507} \code{118,792} \code{12,059} \code{12,099} \code{12,163} \code{12,277} \code{12,451} \code{12,480} \code{12,531} \code{12,552} \code{12,579} \code{12,661} \code{12,700} \code{12,865} \code{12,938} \code{12,979} \code{124,908} \code{125,184} \code{125,937} \code{128,202} \code{13,168} \code{13,176} \code{13,269} \code{13,308} \code{13,315} \code{13,569} \code{13,662} \code{13,677} \code{13,788} \code{13,805} \code{13,952} \code{13,969} \code{131,754} \code{134,738} \code{135,238} \code{135,241} \code{136,085} \code{137,222} \code{139,456} \code{14,017} \code{14,033} \code{14,091} \code{14,163} \code{14,200} \code{14,375} \code{14,446} \code{14,473} \code{14,593} \code{14,697} \code{14,736} \code{14,749} \code{14,804} \code{14,813} \code{14,870} \code{141,409} \code{143,018} \code{143,806} \code{149,266} \code{15,273} \code{15,314} \code{15,441} \code{15,468} \code{15,644} \code{15,651} \code{15,652} \code{15,723} \code{15,725} \code{15,846} \code{15,951} \code{150,490} \code{152,237} \code{154,183} \code{155,667} \code{158,382} \code{159,906} \code{16,050} \code{16,154} \code{16,193} \code{16,202} \code{16,440} \code{16,450} \code{16,531} \code{16,684} \code{16,747} \code{16,757} \code{16,826} \code{16,833} \code{16,855} \code{16,931} \code{161,236} \code{167,660} \code{17,025} \code{17,063} \code{17,150} \code{17,240} \code{17,281} \code{17,289} \code{17,315} \code{17,367} \code{17,373} \code{17,410} \code{17,447} \code{17,647} \code{17,771} \code{17,863} \code{17,976} \code{171,404} \code{18,014} \code{18,111} \code{18,117} \code{18,196} \code{18,414} \code{18,489} \code{18,547} \code{18,619} \code{18,921} \code{180,242} \code{183,234} \code{183,239} \code{184,443} \code{186,357} \code{186,662} \code{186,807} \code{187,952} \code{19,026} \code{19,057} \code{19,668} \code{19,874} \code{19,920} \code{190,293} \code{191,882} \code{193,259} \code{195,936} \code{196,138} \code{196,730} \code{197,965} \code{2,004} \code{2,005} \code{2,012} \code{2,036} \code{2,051} \code{2,055} \code{2,058} \code{2,069} \code{2,087} \code{2,091} \code{2,101} \code{2,106} \code{2,114,888} \code{2,116} \code{2,139} \code{2,140,568} \code{2,150} \code{2,170} \code{2,181} \code{2,203} \code{2,221} \code{2,221,136} \code{2,239} \code{2,254} \code{2,259} \code{2,276} \code{2,295} \code{2,307} \code{2,311} \code{2,323} \code{2,349} \code{2,366} \code{2,373,924} \code{2,401} \code{2,404} \code{2,405} \code{2,456,985} \code{2,481} \code{2,501} \code{2,517} \code{2,525} \code{2,536} \code{2,552} \code{2,557} \code{2,562} \code{2,573} \code{2,577} \code{2,596} \code{2,600} \code{2,648} \code{2,677} \code{2,716} \code{2,744} \code{2,750} \code{2,758} \code{2,766} \code{2,787} \code{2,818} \code{2,820} \code{2,831} \code{2,867} \code{2,876} \code{2,878} \code{2,891} \code{2,902} \code{2,916} \code{2,918} \code{2,945} \code{2,953} \code{2,958} \code{2,962} \code{2,963} \code{2,970} \code{2,972} \code{2,977} \code{2,992} \code{20,024} \code{20,039} \code{20,056} \code{20,150} \code{20,286} \code{20,503} \code{20,724} \code{20,725} \code{20,761} \code{20,808} \code{20,927} \code{203,112} \code{207,530} \code{21,151} \code{21,329} \code{21,421} \code{21,698} \code{21,962} \code{213,479} \code{22,076} \code{22,387} \code{22,845} \code{229,974} \code{23,139} \code{23,216} \code{23,234} \code{23,371} \code{23,484} \code{23,546} \code{236,122} \code{24,106} \code{24,181} \code{24,664} \code{243,557} \code{247,728} \code{248,770} \code{25,136} \code{25,202} \code{25,275} \code{25,974} \code{250,776} \code{250,958} \code{253,564} \code{258,234} \code{26,045,385} \code{26,151} \code{26,202} \code{26,538} \code{26,632} \code{26,640} \code{26,790} \code{26,916} \code{263,417} \code{269,945} \code{27,540} \code{27,670} \code{27,778} \code{27,841} \code{27,850} \code{277,789} \code{28,433} \code{28,443} \code{28,764} \code{28,916} \code{28,994} \code{280,637} \code{283,149} \code{284,012} \code{284,165} \code{29,023} \code{29,079} \code{29,302} \code{291,749} \code{297,696} \code{3,021} \code{3,026} \code{3,049} \code{3,056} \code{3,066} \code{3,119} \code{3,121} \code{3,136} \code{3,164} \code{3,189} \code{3,199} \code{3,201} \code{3,213} \code{3,229} \code{3,266} \code{3,354} \code{3,360} \code{3,401} \code{3,409} \code{3,430} \code{3,437} \code{3,438} \code{3,439} \code{3,444} \code{3,446} \code{3,469} \code{3,490} \code{3,510} \code{3,527} \code{3,528} \code{3,575} \code{3,589} \code{3,591} \code{3,602} \code{3,605} \code{3,613} \code{3,658} \code{3,683} \code{3,684} \code{3,695} \code{3,699} \code{3,727} \code{3,732} \code{3,739} \code{3,746} \code{3,752} \code{3,755} \code{3,772} \code{3,773} \code{3,781} \code{3,786} \code{3,809} \code{3,823} \code{3,874} \code{3,888} \code{3,900} \code{3,917} \code{3,925} \code{3,926} \code{3,986} \code{3,990} \code{30,017} \code{30,192} \code{30,548} \code{300,213} \code{302,576} \code{303,833} \code{307,173} \code{315,243} \code{32,162} \code{32,228} \code{32,436} \code{32,698} \code{33,196} \code{33,383} \code{33,681} \code{33,689} \code{33,967} \code{34,212} \code{34,388} \code{34,494} \code{34,656} \code{343,592} \code{35,057} \code{35,225} \code{35,523} \code{35,547} \code{35,568} \code{36,095} \code{36,167} \code{36,334} \code{36,361} \code{36,756} \code{36,775} \code{38,284} \code{38,475} \code{38,518} \code{39,136} \code{39,346} \code{392,133} \code{398} \code{4,091} \code{4,097} \code{4,098} \code{4,102} \code{4,105} \code{4,155} \code{4,156} \code{4,186} \code{4,223} \code{4,225} \code{4,227} \code{4,236} \code{4,240} \code{4,275} \code{4,289} \code{4,301} \code{4,311} \code{4,330} \code{4,340} \code{4,403} \code{4,433} \code{4,463} \code{4,477} \code{4,489} \code{4,532} \code{4,539} \code{4,567} \code{4,576} \code{4,588} \code{4,591,122} \code{4,601} \code{4,652} \code{4,666} \code{4,681} \code{4,686} \code{4,697} \code{4,718} \code{4,779} \code{4,795} \code{4,813} \code{4,843} \code{4,898} \code{4,908} \code{4,957} \code{4,959} \code{40,418} \code{40,997} \code{41,302} \code{41,484} \code{41,525} \code{418,925} \code{42,008} \code{43,076} \code{43,127} \code{43,665} \code{43,870} \code{436,599} \code{436,759} \code{437} \code{44,207} \code{44,313} \code{44,422} \code{44,478} \code{44,703} \code{44,901} \code{44,953} \code{45,068} \code{45,511} \code{450,512} \code{459} \code{46,107} \code{46,225} \code{46,254} \code{46,287} \code{46,302} \code{46,426} \code{46,775} \code{469,616} \code{47,222} \code{47,601} \code{47,933} \code{48,778} \code{48,917} \code{49,334} \code{49,349} \code{495,102} \code{497} \code{5,019} \code{5,026} \code{5,043} \code{5,060} \code{5,096} \code{5,102} \code{5,129} \code{5,178} \code{5,200} \code{5,213} \code{5,234} \code{5,331} \code{5,342,979} \code{5,351} \code{5,378} \code{5,434} \code{5,438} \code{5,510} \code{5,545} \code{5,546} \code{5,555} \code{5,564} \code{5,565} \code{5,570} \code{5,577} \code{5,635} \code{5,640} \code{5,709} \code{5,778} \code{5,802} \code{5,834} \code{5,836} \code{5,843} \code{5,890} \code{5,894} \code{5,904} \code{5,917} \code{5,938} \code{5,975} \code{50,163} \code{50,302} \code{50,873} \code{50,999} \code{51,551} \code{52,588} \code{52,930} \code{520,084} \code{53,398} \code{53,515} \code{53,844} \code{54,815} \code{54,819} \code{54,834} \code{544} \code{55,851} \code{550,775} \code{56,838} \code{56,927} \code{57,445} \code{57,877} \code{573} \code{58,160} \code{58,257} \code{580} \code{581,153} \code{582} \code{589,974} \code{6,018} \code{6,053} \code{6,105} \code{6,114} \code{6,120} \code{6,128} \code{6,249} \code{6,292} \code{6,305} \code{6,360} \code{6,416} \code{6,431} \code{6,471} \code{6,506} \code{6,567} \code{6,716} \code{6,745} \code{6,769} \code{6,770} \code{6,774} \code{6,805} \code{6,807} \code{6,810} \code{6,852} \code{6,873} \code{6,884} \code{6,888} \code{6,906} \code{60,412} \code{61,276} \code{613} \code{62,580} \code{62,945} \code{623,382} \code{634} \code{64,686} \code{640} \code{65,232} \code{656} \code{657,441} \code{66,773} \code{67,128} \code{67,674} \code{68,646} \code{69,079} \code{69,219} \code{69,382} \code{69,519} \code{69,821} \code{69,979} \code{692} \code{697} \code{7,097} \code{7,131} \code{7,140} \code{7,148} \code{7,179} \code{7,210} \code{7,244} \code{7,254} \code{7,316} \code{7,323} \code{7,435} \code{7,456} \code{7,461} \code{7,496} \code{7,537} \code{7,546} \code{7,570} \code{7,622} \code{7,692} \code{7,693} \code{7,728} \code{7,739} \code{7,744} \code{7,765} \code{7,766} \code{7,790} \code{7,820} \code{7,823} \code{7,832} \code{7,852} \code{7,863} \code{7,924} \code{7,983} \code{7,992} \code{72,302} \code{73,076} \code{73,871} \code{732} \code{735} \code{74,619} \code{74,660} \code{74,675} \code{74,683} \code{743,426} \code{747,747} \code{75,200} \code{75,368} \code{75,579} \code{75,979} \code{753} \code{754} \code{76,287} \code{76,574} \code{77,540} \code{77,594} \code{77,856} \code{775} \code{777} \code{788} \code{79,059} \code{79,393} \code{797,570} \code{8,054} \code{8,069} \code{8,100} \code{8,146} \code{8,216} \code{8,291} \code{8,318} \code{8,346} \code{8,377} \code{8,382} \code{8,493} \code{8,575} \code{8,598} \code{8,663} \code{8,706} \code{8,749} \code{8,772} \code{8,783} \code{8,801} \code{8,863} \code{8,924} \code{8,947} \code{80,858} \code{801} \code{807} \code{808} \code{814,784} \code{829} \code{838} \code{839} \code{84,326} \code{840} \code{844} \code{86,108} \code{86,430} \code{866} \code{868} \code{87,394} \code{870} \code{873} \code{875} \code{88,740} \code{884} \code{89,201} \code{89,733} \code{893} \code{9,001} \code{9,025} \code{9,030} \code{9,051} \code{9,109} \code{9,111} \code{9,120} \code{9,121} \code{9,164} \code{9,186} \code{9,203} \code{9,270} \code{9,395} \code{9,440} \code{9,444} \code{9,514} \code{9,548} \code{9,647} \code{9,708} \code{9,718} \code{9,740} \code{9,747} \code{9,919} \code{90,285} \code{903} \code{908} \code{91,072} \code{91,446} \code{92,998} \code{933} \code{940} \code{949} \code{95,211} \code{950} \code{96,676} \code{967} \code{97,664} \code{976} \code{977} \code{982,065} \code{983} \code{99,349} \code{993} \code{-12} \code{-21} \code{-28} \code{1,005} \code{1,007} \code{1,009,619} \code{1,026} \code{1,030} \code{1,041} \code{1,044} \code{1,050} \code{1,069} \code{1,075,579} \code{1,077} \code{1,095} \code{1,096} \code{1,097} \code{1,107} \code{1,121,070} \code{1,123} \code{1,125} \code{1,126} \code{1,140,617} \code{1,163} \code{1,188} \code{1,193,438} \code{1,200} \code{1,226} \code{1,227} \code{1,244} \code{1,268} \code{1,270} \code{1,272} \code{1,278} \code{1,290} \code{1,294} \code{1,305} \code{1,306} \code{1,309} \code{1,318} \code{1,320} \code{1,327,585} \code{1,336} \code{1,349} \code{1,351} \code{1,358} \code{1,387} \code{1,397} \code{1,418} \code{1,420} \code{1,423} \code{1,423,135} \code{1,442} \code{1,447} \code{1,452,132} \code{1,467} \code{1,468,947} \code{1,469} \code{1,481} \code{1,507} \code{1,520} \code{1,534,760} \code{1,540} \code{1,551} \code{1,561} \code{1,564} \code{1,567} \code{1,602} \code{1,610} \code{1,618} \code{1,620,218} \code{1,624} \code{1,626} \code{1,664} \code{1,674} \code{1,685} \code{1,698} \code{1,711} \code{1,715} \code{1,718} \code{1,731} \code{1,732} \code{1,735,411} \code{1,743} \code{1,775} \code{1,784} \code{1,791} \code{1,794} \code{1,795} \code{1,796} \code{1,798} \code{1,799} \code{1,806} \code{1,819} \code{1,833} \code{1,837} \code{1,847} \code{1,862} \code{1,869} \code{1,905} \code{1,912} \code{1,951} \code{1,958} \code{1,960} \code{1,964} \code{1,974} \code{10,036} \code{10,054} \code{10,104} \code{10,162} \code{10,197} \code{10,230} \code{10,267} \code{10,304} \code{10,347} \code{10,388} \code{10,403} \code{10,455} \code{10,494} \code{10,497} \code{10,519} \code{10,522} \code{10,525} \code{10,610} \code{10,619} \code{10,746} \code{10,848} \code{10,892} \code{10,901} \code{10,902} \code{10,966} \code{10,996} \code{101,974} \code{102,067} \code{102,096} \code{102,633} \code{103,631} \code{103,636} \code{103,888} \code{105,893} \code{108,071} \code{108,450} \code{11,018} \code{11,157} \code{11,176} \code{11,215} \code{11,259} \code{11,372} \code{11,376} \code{11,506} \code{11,521} \code{11,557} \code{11,615} \code{11,676} \code{11,677} \code{11,723} \code{11,812} \code{11,877} \code{11,913} \code{11,982} \code{11,998} \code{110,294} \code{110,956} \code{112,895} \code{113,107} \code{115,675} \code{116,582} \code{119,843} \code{12,037} \code{12,170} \code{12,201} \code{12,250} \code{12,267} \code{12,411} \code{12,456} \code{12,493} \code{12,524} \code{12,538} \code{12,539} \code{12,565} \code{12,628} \code{12,732} \code{12,789} \code{12,876} \code{12,986} \code{120,513} \code{122,769} \code{123,650} \code{123,875} \code{125,833} \code{128,153} \code{128,654} \code{129,275} \code{129,639} \code{13,042} \code{13,046} \code{13,052} \code{13,076} \code{13,090} \code{13,257} \code{13,273} \code{13,277} \code{13,288} \code{13,577} \code{13,602} \code{13,821} \code{13,825} \code{13,921} \code{13,970} \code{13,997} \code{130,974} \code{132,255} \code{132,546} \code{135,549} \code{14,072} \code{14,083} \code{14,159} \code{14,233} \code{14,262} \code{14,309} \code{14,338} \code{14,371} \code{14,729} \code{14,750} \code{14,803} \code{14,826} \code{14,886} \code{14,931} \code{14,932} \code{14,968} \code{140,883} \code{142,364} \code{145,272} \code{15,002} \code{15,151} \code{15,205} \code{15,266} \code{15,284} \code{15,369} \code{15,404} \code{15,500} \code{15,562} \code{15,793} \code{15,870} \code{151,056} \code{152,252} \code{157,381} \code{157,832} \code{16,013} \code{16,023} \code{16,098} \code{16,149} \code{16,152} \code{16,208} \code{16,242} \code{16,456} \code{16,563} \code{16,657} \code{16,827} \code{16,915} \code{16,925} \code{160,337} \code{160,558} \code{160,864} \code{162,273} \code{163,392} \code{164,896} \code{167,892} \code{17,108} \code{17,127} \code{17,212} \code{17,306} \code{17,330} \code{17,404} \code{17,423} \code{17,494} \code{17,534} \code{17,683} \code{17,836} \code{17,890} \code{171,227} \code{172,361} \code{173,153} \code{173,502} \code{174,937} \code{18,265} \code{18,354} \code{18,441} \code{18,482} \code{18,504} \code{18,530} \code{18,638} \code{18,668} \code{18,721} \code{18,730} \code{18,817} \code{18,844} \code{182,082} \code{183,115} \code{183,731} \code{19,053} \code{19,110} \code{19,158} \code{19,307} \code{19,316} \code{19,592} \code{19,654} \code{19,678} \code{19,686} \code{19,831} \code{19,849} \code{19,915} \code{19,979} \code{193,453} \code{194,603} \code{195,898} \code{2,010} \code{2,030} \code{2,042} \code{2,053} \code{2,072} \code{2,096} \code{2,108} \code{2,112} \code{2,155} \code{2,160} \code{2,176,999} \code{2,205} \code{2,208} \code{2,211} \code{2,213} \code{2,225} \code{2,253} \code{2,255} \code{2,258} \code{2,266,371} \code{2,315} \code{2,316} \code{2,329} \code{2,343,482} \code{2,376} \code{2,383} \code{2,393} \code{2,410} \code{2,413} \code{2,430,920} \code{2,434} \code{2,461} \code{2,463,960} \code{2,521} \code{2,526} \code{2,540} \code{2,550} \code{2,558,358} \code{2,576} \code{2,585} \code{2,595} \code{2,598} \code{2,603} \code{2,623,552} \code{2,644} \code{2,650} \code{2,669} \code{2,697} \code{2,700} \code{2,702} \code{2,709} \code{2,712} \code{2,722} \code{2,728} \code{2,788} \code{2,790} \code{2,795} \code{2,796} \code{2,797} \code{2,821} \code{2,824} \code{2,849} \code{2,853} \code{2,855} \code{2,861} \code{2,871} \code{2,884} \code{2,914} \code{2,951} \code{2,954} \code{2,957} \code{2,985} \code{2,986} \code{2,996} \code{2,998} \code{20,060} \code{20,106} \code{20,130} \code{20,185} \code{20,364} \code{20,416} \code{20,441} \code{20,485} \code{20,551} \code{20,567} \code{20,844} \code{200,252} \code{202,184} \code{206,929} \code{208,050} \code{208,102} \code{21,057} \code{21,166} \code{21,239} \code{21,316} \code{21,353} \code{21,450} \code{21,650} \code{21,805} \code{21,870} \code{21,911} \code{216,014} \code{22,128} \code{22,376} \code{22,560} \code{22,846} \code{220,461} \code{220,681} \code{220,724} \code{220,753} \code{226,380} \code{226,760} \code{23,097} \code{23,157} \code{23,178} \code{23,246} \code{23,318} \code{23,376} \code{23,405} \code{23,445} \code{23,479} \code{23,672} \code{23,689} \code{23,776} \code{23,938} \code{234,065} \code{235,906} \code{237,419} \code{238,616} \code{24,252} \code{24,371} \code{24,495} \code{24,499} \code{24,524} \code{24,683} \code{24,709} \code{24,764} \code{24,921} \code{24,935} \code{243,553} \code{244,902} \code{248,625} \code{25,111} \code{25,214} \code{25,240} \code{25,278} \code{25,459} \code{25,490} \code{25,492} \code{25,828} \code{250,527} \code{251,884} \code{252,263} \code{26,069} \code{26,081} \code{26,290} \code{26,411} \code{26,561} \code{26,569} \code{26,674} \code{26,681} \code{26,726} \code{26,987} \code{264,243} \code{268,689} \code{27,033} \code{27,100} \code{27,527} \code{27,987} \code{270,331} \code{278,810} \code{28,124} \code{28,460} \code{28,598} \code{28,754} \code{280,141} \code{281,061} \code{281,594} \code{282,094} \code{287,359} \code{288,324} \code{29,130} \code{29,598} \code{29,725} \code{3,018} \code{3,054} \code{3,061} \code{3,086} \code{3,111} \code{3,112} \code{3,127} \code{3,137} \code{3,140,509} \code{3,194} \code{3,200} \code{3,210} \code{3,218} \code{3,230,085} \code{3,249} \code{3,256} \code{3,261} \code{3,312} \code{3,314} \code{3,333} \code{3,344} \code{3,352} \code{3,362} \code{3,388} \code{3,456} \code{3,495} \code{3,516} \code{3,555} \code{3,560} \code{3,562} \code{3,587} \code{3,600} \code{3,638} \code{3,644} \code{3,649} \code{3,692} \code{3,713} \code{3,718} \code{3,757} \code{3,763} \code{3,789} \code{3,820} \code{3,836} \code{3,838} \code{3,862} \code{3,884} \code{3,897} \code{3,924} \code{3,928} \code{3,931} \code{3,948} \code{3,983} \code{3,996} \code{30,676} \code{304,036} \code{308,536} \code{31,010} \code{316,113} \code{32,015} \code{32,199} \code{32,891} \code{32,971} \code{324,365} \code{324,382} \code{324,791} \code{329,852} \code{33,145} \code{33,161,440} \code{33,308} \code{33,429} \code{33,978} \code{34,633} \code{34,705} \code{34,926} \code{348,241} \code{35,094} \code{35,451} \code{35,715} \code{35,914} \code{35,943} \code{35,968} \code{352,400} \code{357,692} \code{359,758} \code{36,015} \code{36,694} \code{36,702} \code{37,050} \code{37,109} \code{37,303} \code{37,999} \code{371,666} \code{375,542} \code{377,822} \code{377,949} \code{38,176} \code{38,566} \code{38,571} \code{38,719} \code{38,817} \code{38,893} \code{38,920} \code{386} \code{39,572} \code{39,843} \code{39,857} \code{39,864} \code{394,558} \code{395,711} \code{397,295} \code{4,003} \code{4,013} \code{4,025} \code{4,044} \code{4,067} \code{4,079} \code{4,087} \code{4,108} \code{4,118} \code{4,124} \code{4,136} \code{4,167} \code{4,180} \code{4,206} \code{4,218} \code{4,219} \code{4,233} \code{4,245} \code{4,252} \code{4,271} \code{4,274} \code{4,310} \code{4,352} \code{4,354} \code{4,388} \code{4,397} \code{4,402} \code{4,419} \code{4,443} \code{4,448} \code{4,482} \code{4,496} \code{4,520} \code{4,581} \code{4,599} \code{4,611} \code{4,620} \code{4,677} \code{4,684} \code{4,704} \code{4,729} \code{4,743} \code{4,748} \code{4,754} \code{4,782} \code{4,822} \code{4,823} \code{4,836} \code{4,875} \code{4,879} \code{4,890} \code{4,891} \code{4,917} \code{4,927} \code{4,928} \code{4,931} \code{4,946} \code{4,953} \code{4,969} \code{4,970} \code{4,980} \code{4,986} \code{40,367} \code{40,616} \code{41,047} \code{41,517} \code{417,048} \code{42,321} \code{42,687} \code{42,757} \code{42,874} \code{425,560} \code{434,670} \code{448,661} \code{45,454} \code{46,040} \code{46,504} \code{46,815} \code{47,164} \code{47,590} \code{47,764} \code{47,787} \code{47,908} \code{477,889} \code{48,143} \code{48,441} \code{49,231} \code{49,459} \code{5,001} \code{5,007} \code{5,027} \code{5,031} \code{5,059} \code{5,075} \code{5,094} \code{5,109} \code{5,115} \code{5,202} \code{5,228} \code{5,237} \code{5,253} \code{5,269} \code{5,296} \code{5,319} \code{5,347} \code{5,349} \code{5,394} \code{5,436} \code{5,468} \code{5,486} \code{5,487} \code{5,517} \code{5,541} \code{5,571} \code{5,579} \code{5,599} \code{5,630} \code{5,647} \code{5,694} \code{5,734} \code{5,747} \code{5,786} \code{5,806} \code{5,807} \code{5,828} \code{5,929} \code{5,972,356} \code{50,573} \code{51,486} \code{52,164} \code{52,440} \code{52,630} \code{52,731} \code{52,917} \code{53,178} \code{54,236} \code{54,878} \code{55,535} \code{558,632} \code{56,294} \code{56,773} \code{56,984} \code{57,033} \code{58,173} \code{58,288} \code{58,315} \code{58,602} \code{580,552} \code{584,297} \code{585,567} \code{59,094} \code{59,371} \code{59,934} \code{6,006} \code{6,055} \code{6,095} \code{6,096} \code{6,132} \code{6,145} \code{6,148} \code{6,189} \code{6,190} \code{6,201} \code{6,215} \code{6,233} \code{6,372} \code{6,385} \code{6,387} \code{6,442,416} \code{6,461} \code{6,485} \code{6,496} \code{6,529} \code{6,561} \code{6,601} \code{6,629} \code{6,648} \code{6,686} \code{6,722} \code{6,733} \code{6,781} \code{6,785} \code{6,787} \code{6,800} \code{6,877} \code{6,882} \code{6,896} \code{6,899} \code{6,938} \code{6,956} \code{6,960} \code{60,102} \code{60,258} \code{60,749} \code{603,080} \code{61,266} \code{61,779} \code{61,895} \code{622} \code{63,236} \code{63,336} \code{63,511} \code{63,578} \code{631,565} \code{64,047} \code{64,118} \code{64,410} \code{64,749} \code{64,808} \code{64,837} \code{64,871} \code{65,902} \code{65,952} \code{66,520} \code{66,711} \code{66,893} \code{66,946} \code{665} \code{666,658} \code{67,964} \code{671} \code{676,551} \code{68,911} \code{69,356} \code{69,757} \code{696} \code{7,054} \code{7,070} \code{7,094} \code{7,105} \code{7,111} \code{7,121} \code{7,133} \code{7,151} \code{7,156} \code{7,164} \code{7,188} \code{7,190} \code{7,297} \code{7,344} \code{7,365} \code{7,451} \code{7,498} \code{7,534} \code{7,552} \code{7,562} \code{7,582} \code{7,699} \code{7,725} \code{7,805} \code{7,828} \code{7,859} \code{7,914} \code{7,915} \code{7,927} \code{70,637} \code{70,790} \code{708} \code{71,254} \code{71,975} \code{711} \code{712,884} \code{719} \code{73,549} \code{730} \code{738} \code{74,448} \code{75,187} \code{75,314} \code{75,850} \code{752} \code{76,676} \code{76,870} \code{76,981} \code{77,145} \code{77,360} \code{77,970} \code{78,456} \code{79,487} \code{793} \code{797,608} \code{8,043} \code{8,092} \code{8,101} \code{8,104} \code{8,113} \code{8,181} \code{8,207} \code{8,254} \code{8,405} \code{8,410} \code{8,472} \code{8,505} \code{8,529} \code{8,545} \code{8,683} \code{8,704} \code{8,792} \code{8,816} \code{8,900} \code{80,606} \code{81,550} \code{818} \code{82,194} \code{834,510} \code{84,905} \code{848} \code{85,069} \code{85,652} \code{863} \code{87,518} \code{87,851} \code{87,936} \code{871} \code{876} \code{88,407} \code{88,496} \code{89,031} \code{89,199} \code{890} \code{896,372} \code{9,005} \code{9,041} \code{9,046} \code{9,057} \code{9,113} \code{9,159} \code{9,194} \code{9,237} \code{9,362} \code{9,366} \code{9,460} \code{9,465} \code{9,481} \code{9,489} \code{9,493} \code{9,494} \code{9,509} \code{9,710} \code{9,743} \code{9,746} \code{9,775} \code{9,781} \code{910} \code{915} \code{917} \code{919} \code{926} \code{93,465} \code{938} \code{96,223} \code{96,237} \code{961} \code{97,766} \code{97,850} \code{973} \code{98,653} \code{99,775} \code{994} \code{-20} \code{-36} \code{-8} \code{1,010} \code{1,011} \code{1,082,402} \code{1,102} \code{1,141} \code{1,142,380} \code{1,144} \code{1,146} \code{1,149} \code{1,151} \code{1,182} \code{1,183} \code{1,187} \code{1,191} \code{1,198} \code{1,210,673} \code{1,212} \code{1,222} \code{1,223} \code{1,230} \code{1,242,179} \code{1,243} \code{1,245} \code{1,252} \code{1,255} \code{1,287} \code{1,296} \code{1,330} \code{1,339} \code{1,341} \code{1,342} \code{1,360,296} \code{1,363} \code{1,380} \code{1,386} \code{1,390} \code{1,398} \code{1,405} \code{1,408} \code{1,409} \code{1,425} \code{1,462} \code{1,465} \code{1,468} \code{1,471} \code{1,493,231} \code{1,509} \code{1,514} \code{1,518} \code{1,536} \code{1,539} \code{1,541} \code{1,552,154} \code{1,564,798} \code{1,566} \code{1,578,568} \code{1,593} \code{1,637} \code{1,657} \code{1,671} \code{1,672} \code{1,687} \code{1,688,807} \code{1,692} \code{1,701} \code{1,704} \code{1,720} \code{1,739} \code{1,753} \code{1,804} \code{1,809} \code{1,813} \code{1,825} \code{1,834,473} \code{1,911} \code{1,914} \code{1,917} \code{1,923} \code{1,955} \code{1,959} \code{1,970,468} \code{1,971} \code{10,004} \code{10,010} \code{10,035} \code{10,101} \code{10,106} \code{10,128} \code{10,158} \code{10,196} \code{10,211} \code{10,343} \code{10,581} \code{10,601} \code{10,759} \code{10,764} \code{10,777} \code{10,929} \code{10,934} \code{100,205} \code{103,286} \code{103,801} \code{104,288} \code{104,332} \code{104,390} \code{106,411} \code{109,731} \code{11,089} \code{11,092} \code{11,119} \code{11,152} \code{11,230} \code{11,234} \code{11,263} \code{11,264} \code{11,295} \code{11,306} \code{11,314} \code{11,393} \code{11,394} \code{11,424} \code{11,463} \code{11,482} \code{11,579} \code{11,585} \code{11,705} \code{11,735} \code{11,788} \code{11,796} \code{11,844} \code{11,868} \code{11,878} \code{11,978} \code{11,988} \code{11,992} \code{110,277} \code{110,686} \code{110,814} \code{111,091} \code{111,368} \code{111,799} \code{112,057} \code{113,910} \code{117,927} \code{118,054} \code{118,154} \code{119,035} \code{119,963} \code{12,006} \code{12,009} \code{12,078} \code{12,205} \code{12,380} \code{12,400} \code{12,402} \code{12,415} \code{12,432} \code{12,526} \code{12,592} \code{12,595} \code{12,633} \code{12,697} \code{12,774} \code{12,795} \code{12,852} \code{12,934} \code{12,935} \code{121,276} \code{122,126} \code{127,849} \code{13,184} \code{13,194} \code{13,408} \code{13,473} \code{13,586} \code{13,617} \code{13,763} \code{13,776} \code{13,793} \code{13,803} \code{13,826} \code{13,860} \code{13,906} \code{13,996} \code{130,093} \code{131,459} \code{131,854} \code{132,660} \code{133,447} \code{135,019} \code{135,032} \code{137,362} \code{137,567} \code{139,555} \code{14,004} \code{14,025} \code{14,146} \code{14,162} \code{14,210} \code{14,240} \code{14,292} \code{14,300} \code{14,340} \code{14,353} \code{14,461} \code{14,492} \code{14,564} \code{14,695} \code{14,700} \code{14,784} \code{14,843} \code{14,917} \code{14,996} \code{142,311} \code{142,493} \code{143,340} \code{145,953} \code{146,413} \code{15,046} \code{15,048} \code{15,108} \code{15,187} \code{15,354} \code{15,438} \code{15,501} \code{15,610} \code{15,687} \code{15,742} \code{15,753} \code{15,830} \code{15,888} \code{15,895} \code{15,911} \code{153,563} \code{154,627} \code{154,645} \code{156,708} \code{158,606} \code{16,071} \code{16,072} \code{16,125} \code{16,165} \code{16,194} \code{16,240} \code{16,262} \code{16,407} \code{16,479} \code{16,595} \code{16,823} \code{16,842} \code{16,879} \code{16,958} \code{160,989} \code{161,297} \code{162,825} \code{165,721} \code{17,070} \code{17,100} \code{17,138} \code{17,244} \code{17,504} \code{17,524} \code{17,527} \code{17,553} \code{17,576} \code{17,681} \code{17,696} \code{17,773} \code{17,886} \code{17,935} \code{170,165} \code{171,019} \code{171,058} \code{173,759} \code{173,908} \code{176,004} \code{177,138} \code{178,417} \code{18,075} \code{18,349} \code{18,351} \code{18,366} \code{18,775} \code{180,613} \code{183,388} \code{188,807} \code{189,317} \code{19,135} \code{19,242} \code{19,373} \code{19,546} \code{19,648} \code{19,657} \code{19,893} \code{19,969} \code{190,592} \code{192,530} \code{199,503} \code{199,564} \code{2,006} \code{2,017} \code{2,018} \code{2,020} \code{2,024} \code{2,025} \code{2,054} \code{2,066} \code{2,077} \code{2,089} \code{2,097} \code{2,100} \code{2,110} \code{2,119} \code{2,122} \code{2,125} \code{2,134} \code{2,142,022} \code{2,143} \code{2,158} \code{2,167} \code{2,183} \code{2,194} \code{2,198} \code{2,227} \code{2,236} \code{2,314} \code{2,348} \code{2,370} \code{2,384} \code{2,386} \code{2,408} \code{2,411} \code{2,417} \code{2,417,571} \code{2,424} \code{2,426} \code{2,442} \code{2,455} \code{2,477} \code{2,487} \code{2,497} \code{2,499} \code{2,516} \code{2,523,566} \code{2,528} \code{2,570} \code{2,578} \code{2,591} \code{2,608} \code{2,611} \code{2,624,809} \code{2,631} \code{2,632} \code{2,680} \code{2,681} \code{2,682} \code{2,703} \code{2,704} \code{2,740} \code{2,743} \code{2,754} \code{2,774} \code{2,779} \code{2,791} \code{2,800} \code{2,807} \code{2,809} \code{2,882,607} \code{2,928} \code{2,930} \code{2,936} \code{2,964,995} \code{20,098} \code{20,134} \code{20,234} \code{20,256} \code{20,274} \code{20,295} \code{20,298} \code{20,434} \code{20,447} \code{20,533} \code{20,702} \code{20,751} \code{20,793} \code{20,827} \code{20,977} \code{20,978} \code{20,990} \code{204,595} \code{208,283} \code{21,004} \code{21,321} \code{21,338} \code{21,385} \code{21,551} \code{21,600} \code{21,635} \code{21,756} \code{21,815} \code{21,847} \code{21,943} \code{21,998} \code{211,774} \code{213,659} \code{216,337} \code{22,006} \code{22,125} \code{22,138} \code{22,164} \code{22,230} \code{22,343} \code{22,363} \code{22,644} \code{22,851} \code{22,953} \code{221,926} \code{222,570} \code{224,535} \code{23,010} \code{23,268} \code{23,286} \code{23,458} \code{23,480} \code{23,578} \code{23,641} \code{23,845} \code{235,059} \code{238,917} \code{24,016} \code{24,018} \code{24,162} \code{24,263} \code{24,266} \code{24,660} \code{24,740} \code{24,812} \code{24,816} \code{24,836} \code{240,170} \code{247,809} \code{248,785} \code{25,086} \code{25,209} \code{25,305} \code{25,484} \code{251,098} \code{252,198} \code{255,115} \code{256,140} \code{256,393} \code{26,065} \code{26,104} \code{26,166} \code{26,510} \code{26,605} \code{26,619} \code{26,634} \code{26,685} \code{26,697} \code{26,762} \code{26,808} \code{264,325} \code{265,025} \code{268,468} \code{268,792} \code{27,153} \code{27,391} \code{27,425} \code{27,653} \code{27,682} \code{27,773} \code{27,837} \code{27,958} \code{272,662} \code{272,968} \code{28,388} \code{28,633} \code{28,731} \code{28,736} \code{28,761} \code{28,997} \code{287,675} \code{29,426} \code{29,435} \code{29,438} \code{29,975} \code{293,688} \code{3,005} \code{3,012} \code{3,016} \code{3,020} \code{3,109,369} \code{3,177} \code{3,178} \code{3,196} \code{3,209} \code{3,214} \code{3,215} \code{3,223} \code{3,227} \code{3,228,198} \code{3,243} \code{3,247} \code{3,278} \code{3,281} \code{3,298} \code{3,299} \code{3,319} \code{3,335} \code{3,347} \code{3,350} \code{3,358} \code{3,367} \code{3,377} \code{3,389} \code{3,393,411} \code{3,399} \code{3,429} \code{3,435} \code{3,468} \code{3,470} \code{3,501} \code{3,507} \code{3,509} \code{3,520} \code{3,549} \code{3,620} \code{3,633} \code{3,686} \code{3,695,184} \code{3,719} \code{3,729} \code{3,736} \code{3,751} \code{3,777} \code{3,847} \code{3,889} \code{3,910} \code{3,923} \code{3,933} \code{3,988} \code{30,037} \code{30,167} \code{30,231} \code{30,663} \code{30,802} \code{301,287} \code{31,125} \code{31,145} \code{31,409} \code{31,451} \code{31,922} \code{31,998} \code{311,073} \code{32,061} \code{32,074} \code{32,130} \code{32,315} \code{32,399} \code{32,477} \code{32,568} \code{32,872} \code{33,136} \code{33,323} \code{33,455} \code{33,837} \code{33,871} \code{332,381} \code{335,430} \code{337,600} \code{34,220} \code{34,221} \code{35,108} \code{35,460} \code{35,572} \code{35,602} \code{35,662} \code{35,822} \code{35,903} \code{357,292} \code{359,418} \code{36,046} \code{36,103} \code{36,170} \code{36,672} \code{365,848} \code{369,047} \code{37,117} \code{37,417} \code{37,578} \code{37,774} \code{37,874} \code{37,942} \code{372,887} \code{373,153} \code{373,476} \code{373,606} \code{374,022} \code{38,264} \code{38,341} \code{38,845} \code{38,912} \code{383,824} \code{385,547} \code{39,307,344} \code{39,350} \code{39,625} \code{39,678} \code{39,750} \code{39,809} \code{395,806} \code{4,026} \code{4,053} \code{4,131} \code{4,159} \code{4,174} \code{4,176} \code{4,191} \code{4,193} \code{4,198} \code{4,200} \code{4,231} \code{4,262} \code{4,268} \code{4,270} \code{4,300} \code{4,329} \code{4,338} \code{4,365} \code{4,384} \code{4,386} \code{4,411} \code{4,413} \code{4,430} \code{4,451} \code{4,453} \code{4,456} \code{4,472} \code{4,494} \code{4,502} \code{4,535} \code{4,570} \code{4,636} \code{4,665} \code{4,678} \code{4,705} \code{4,747} \code{4,755} \code{4,769} \code{4,826} \code{4,835} \code{4,838} \code{4,858} \code{4,878} \code{4,883} \code{4,903} \code{4,924} \code{4,937} \code{4,954} \code{4,974} \code{4,985} \code{4,991} \code{40,105} \code{40,238} \code{40,289} \code{40,356} \code{40,547} \code{404,423} \code{404,823} \code{406,965} \code{41,768} \code{41,968} \code{417,493} \code{42,045} \code{42,378} \code{42,439} \code{421,683} \code{424,166} \code{43,365} \code{43,652} \code{43,755} \code{44,185} \code{44,972} \code{440,474} \code{443,513} \code{45,340} \code{45,528} \code{45,593} \code{455,516} \code{459,054} \code{46,658} \code{46,964} \code{47,084} \code{47,157} \code{47,822} \code{478,004} \code{481,146} \code{484,361} \code{49,287} \code{491,138} \code{5,022} \code{5,065} \code{5,078} \code{5,127} \code{5,135} \code{5,142} \code{5,152} \code{5,169} \code{5,172} \code{5,184} \code{5,233} \code{5,261} \code{5,292} \code{5,327} \code{5,366} \code{5,372} \code{5,390} \code{5,416} \code{5,423} \code{5,450} \code{5,454} \code{5,471} \code{5,485} \code{5,513} \code{5,532} \code{5,539} \code{5,566} \code{5,587} \code{5,655} \code{5,663} \code{5,698} \code{5,730} \code{5,735} \code{5,769} \code{5,794} \code{5,801} \code{5,810} \code{5,818} \code{5,832} \code{5,875} \code{5,922} \code{5,949} \code{5,966} \code{5,998} \code{50,127} \code{50,310} \code{508,768} \code{51,235} \code{51,891} \code{515,857} \code{52,030} \code{53,141} \code{53,535} \code{53,650} \code{53,771} \code{53,793} \code{53,802} \code{54,124} \code{54,250} \code{54,518} \code{54,820} \code{543,887} \code{544,458} \code{55,163} \code{55,687} \code{56,114} \code{56,943} \code{57,737} \code{573,708} \code{58,370} \code{58,469} \code{58,589} \code{582,342} \code{586,454} \code{59,044} \code{59,858} \code{6,035} \code{6,054} \code{6,086} \code{6,110} \code{6,126} \code{6,168} \code{6,169} \code{6,220} \code{6,271} \code{6,278} \code{6,302} \code{6,323} \code{6,327} \code{6,350} \code{6,373} \code{6,408} \code{6,410} \code{6,414} \code{6,441} \code{6,489} \code{6,499} \code{6,520} \code{6,539} \code{6,541} \code{6,546} \code{6,560} \code{6,577} \code{6,584} \code{6,594} \code{6,698} \code{6,730} \code{6,731} \code{6,756} \code{6,763} \code{6,783} \code{6,795} \code{6,840} \code{6,949} \code{6,951} \code{6,953} \code{6,964} \code{60,077} \code{60,150} \code{60,491} \code{60,673} \code{62,298} \code{62,431} \code{625,224} \code{64,170} \code{64,239} \code{64,705} \code{64,710} \code{64,737} \code{647} \code{65,086} \code{65,726} \code{65,931} \code{66,714} \code{66,984} \code{669,482} \code{67,184} \code{67,864} \code{67,909} \code{68,300} \code{69,830} \code{7,007} \code{7,072} \code{7,096} \code{7,160} \code{7,191,779} \code{7,234} \code{7,252} \code{7,268} \code{7,273} \code{7,304} \code{7,308} \code{7,334} \code{7,357} \code{7,362} \code{7,373} \code{7,379,911} \code{7,425} \code{7,460} \code{7,642} \code{7,690} \code{7,779} \code{7,871} \code{7,903} \code{7,909} \code{7,919} \code{7,952} \code{70,391} \code{71,276} \code{71,277} \code{71,577} \code{72,838} \code{720,008} \code{722} \code{723} \code{73,857} \code{742,704} \code{748} \code{755} \code{76,041} \code{76,134} \code{76,187} \code{771} \code{773,381} \code{778,580} \code{78,780} \code{784} \code{79,029} \code{79,521} \code{79,947} \code{79,995} \code{797} \code{798} \code{8,033} \code{8,067} \code{8,167} \code{8,193} \code{8,199} \code{8,215} \code{8,296} \code{8,422} \code{8,424} \code{8,428} \code{8,437} \code{8,461} \code{8,495} \code{8,515} \code{8,517} \code{8,536} \code{8,548} \code{8,556} \code{8,618} \code{8,711} \code{8,769} \code{8,774} \code{8,926} \code{8,958} \code{8,984} \code{80,641} \code{80,919} \code{81,192} \code{813,311} \code{82,583} \code{82,930} \code{828,691} \code{83,449} \code{85,149} \code{86,022} \code{86,456} \code{86,660} \code{86,749} \code{86,854} \code{863,632} \code{867,903} \code{87,205} \code{87,645} \code{882} \code{889} \code{89,289} \code{89,923} \code{9,072} \code{9,082} \code{9,162} \code{9,184} \code{9,225} \code{9,317} \code{9,391} \code{9,394} \code{9,400} \code{9,409} \code{9,413} \code{9,432} \code{9,461} \code{9,499} \code{9,584} \code{9,607} \code{9,634} \code{9,639} \code{9,676} \code{9,737} \code{9,798} \code{9,805} \code{9,811} \code{9,813} \code{9,826} \code{9,856} \code{9,862} \code{9,872} \code{9,902} \code{9,938} \code{9,984} \code{90,008} \code{90,540} \code{92,173} \code{92,854} \code{94,173} \code{94,649} \code{94,993} \code{947} \code{948} \code{95,121} \code{95,295} \code{957} \code{97,240} \code{97,575} \code{979} \code{98,149} \code{98,599} \code{985} \code{986} \code{995} \code{996}}
    \item{\code{time_o}}{a numeric vector}
    \item{\code{time_d}}{a numeric vector}
    \item{\code{timediff}}{a numeric vector}
    \item{\code{FPIflow2}}{a numeric vector}
    \item{\code{iszero}}{a logical vector}
    \item{\code{correla}}{a numeric vector}
    \item{\code{logtradebyGDP}}{a numeric vector}
    \item{\code{gravity}}{a numeric vector}
    \item{\code{morethangravity}}{a logical vector}
  }
}
\details{
o is origin and d is destination. Correla is the correlation between the GDP growth rates of the origin and destination over the last 10 years. Comlang_off is TRUE if the two share a common official language. Comleg refers to a common legal origin between the two. Flow is trade flow, FPIflow2 is poorly named, it is actually FPI holdings. 
}
\source{
http://cpis.imf.org/
http://www.cepii.fr/CEPII/en/bdd_modele/presentation.asp?id=27
http://econ.worldbank.org/WBSITE/EXTERNAL/EXTDEC/EXTRESEARCH/0,,contentMDK:23407813~pagePK:64214825~piPK:64214943~theSitePK:469382,00.html
}

\examples{
data(FPI2002Data)
## maybe str(FPI2002Data) ; plot(FPI2002Data) ...
}
\keyword{datasets}
