\name{G_Point7}
\alias{G_Point7}
\docType{data}
\title{
The seven points used for Gaussian Hermite Quadrature
}
\description{
This is purely for user's curiousity, modifying this will not affect the workings of the package as it is either hardcoded in or taken as a parameter.
}
\usage{data("G_Point7")}
\format{
  The format is:
 num [1:7] -3.75 -2.37 -1.15 0 1.15 ...
}
\examples{
data(G_Point7)
}
\keyword{datasets}
