\name{GH_Quadrature}
\alias{GH_Quadrature}
\docType{data}
\title{
The seven weights used for Gaussian Hermite Quadrature
}
\description{
This is purely for user's curiousity, modifying this will not affect the workings of the package as it is either hardcoded in or taken as a parameter.
}
\usage{data("GH_Quadrature")}
\format{
  The format is:
 num [1:7] 0.000548 0.030757 0.240123 0.457143 0.240123 ...
}

\examples{
data(GH_Quadrature)
}
\keyword{datasets}
