\name{gRho}
\alias{gRho}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Purely used in rho_0
}
\description{
See above
}
\usage{
gRho(x1, x2, mu1, mu2, s1, s2, rho12, star_rho12, fhat1, fhat2, G_Point7, GH_Quadrature)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x1}{
See rho_0
}
  \item{x2}{
See rho_0
}
  \item{mu1}{
See rho_0
}
  \item{mu2}{
See rho_0
}
  \item{s1}{
See rho_0
}
  \item{s2}{
See rho_0
}
  \item{rho12}{
See rho_0
}
  \item{star_rho12}{
See rho_0
}
  \item{fhat1}{
See rho_0
}
  \item{fhat2}{
See rho_0
}
  \item{G_Point7}{
See rho_0
}
  \item{GH_Quadrature}{
See rho_0
}
}

\value{
Numeric Scalar
}
\references{
no references
}
\author{
Yan-Xia Lin
}

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
